% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_error_conversion.R
\name{abs_to_rel}
\alias{abs_to_rel}
\title{Convert absolute errors to relative errors in ASTR object}
\usage{
abs_to_rel(df)
}
\arguments{
\item{df}{An ASTR object}
}
\value{
An ASTR object with errors converted to relative
}
\description{
Converts all absolute error columns to relative errors (\%).
}
