% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_PbIso_AgeModels.R
\name{age_models}
\alias{age_models}
\alias{pb_iso_age_model}
\alias{stacey_kramers_1975}
\alias{cumming_richards_1975}
\alias{albarede_juteau_1984}
\title{Calculate lead isotope age models}
\usage{
pb_iso_age_model(
  df,
  ratio_206_204 = "206Pb/204Pb",
  ratio_207_204 = "207Pb/204Pb",
  ratio_208_204 = "208Pb/204Pb",
  model = c("SK75", "CR75", "AJ84", "all")
)

stacey_kramers_1975(
  df,
  ratio_206_204 = "206Pb/204Pb",
  ratio_207_204 = "207Pb/204Pb",
  ratio_208_204 = "208Pb/204Pb"
)

cumming_richards_1975(
  df,
  ratio_206_204 = "206Pb/204Pb",
  ratio_207_204 = "207Pb/204Pb",
  ratio_208_204 = NULL
)

albarede_juteau_1984(
  df,
  ratio_206_204 = "206Pb/204Pb",
  ratio_207_204 = "207Pb/204Pb",
  ratio_208_204 = "208Pb/204Pb"
)
}
\arguments{
\item{df}{The data frame from which the age model parameters should be
calculated.}

\item{ratio_206_204}{Name of the column with the 206Pb/204Pb ratio as
character string. Default is \verb{206Pb/204Pb}.}

\item{ratio_207_204}{Name of the column with the 207Pb/204Pb ratio as
character string. Default is \verb{207Pb/204Pb}.}

\item{ratio_208_204}{Name of the column with the 208Pb/204Pb ratio as
character string. Default is \verb{208Pb/204Pb}.}

\item{model}{Character string with the abbreviation of the model to
calculate:
\itemize{
\item \code{SK75} for Stacey & Kramers (1975)
\item \code{CR75} for Cumming & Richards (1975)
\item \code{AJ84} for Albarède & Juteau (1984)
\item \code{all} for all models at once
}}
}
\value{
If \code{df} is an \link[=ASTR]{ASTR object}, the output is an object of the
same type including the ID column, the contextual columns, the lead isotope
ratios used for calculation of the age model parameters, and the calculated
age model parameters. In all other cases, the data frame provided as input
with columns added for the calculated age model parameters.

The used age model is indicated in the column names of the output by the
abbreviations for the models given above. They represent the initials of
the author's last names and the publication year (e. g. \code{SK75} for Stacey &
Kramers 1975).
}
\description{
The functions calculate the age model and their respective mu and kappa
values according to the publications they are named after (see
\emph{References}). \code{\link[=pb_iso_age_model]{pb_iso_age_model()}} provides a wrapper for them and allows
to calculate all age models simultaneously.
}
\details{
See the references for the respective publications of the age models. The
function for the age model of Albarède & Juteau (1984) is based on the
MATLAB-script of F. Albarède (version 2020-11-06). The age model published in
Albarède et al. (2012) should not be used according to F. Albarède and is
therefore not implemented. Instead, he recommends to use the age model
published in Albarède & Juteau (1984).

The ratio of 208Pb/204Pb is not necessary for \link{cumming_richards_1975}. The
function takes it as argument only to be consistent with the input of the
other age model functions. If provided, it will be ignored.
}
\examples{
# creating example data
df <- tibble::tibble(
  `206Pb/204Pb` = runif(5, min = 18, max = 18.5),
  `207Pb/204Pb` = runif(5, min = 15, max = 15.5),
  `208Pb/204Pb` = runif(5, min = 2, max = 2.2)
)

# calculate values for all age models
pb_iso_age_model(df, model = "all")

# calculate values for a specific age model
pb_iso_age_model(df, model = "SK75")
stacey_kramers_1975(df)

}
\references{
Albarède, F. and Juteau, M. (1984) Unscrambling the lead model
ages. Geochimica et Cosmochimica Acta 48(1), pp. 207-212.
\url{https://dx.doi.org/10.1016/0016-7037(84)90364-8}.

Albarède, F., Desaulty, A.-M. and Blichert-Toft, J. (2012) A geological
perspective on the use of Pb isotopes in Archaeometry. Archaeometry 54, pp.
853–867. \url{https://doi.org/10.1111/j.1475-4754.2011.00653.x}.

Cumming, G.L. and Richards, J.R. (1975) Ore lead isotope ratios in a
continuously changing earth. Earth and Planetary Science Letters 28(2), pp.
155–171. \url{https://dx.doi.org/10.1016/0012-821X(75)90223-X}.

Stacey, J.S. and Kramers, J.D. (1975) Approximation of terrestrial lead
isotope evolution by a two-stage model. Earth and Planetary Science Letters
26(2), pp. 207–221. \url{https://dx.doi.org/10.1016/0012-821X(75)90088-6}.
}
\seealso{
Other Pb isotope functions:
\code{\link[=as.ref_data]{as.ref_data()}},
\code{\link[=geom_sk_lines]{geom_sk_lines()}},
\code{\link[=pb_iso_endmembers]{pb_iso_endmembers()}},
\code{\link[=pb_iso_prov_dist]{pb_iso_prov_dist()}},
\code{\link[=pb_iso_train_data]{pb_iso_train_data()}}
}
\concept{Pb isotope functions}
