% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{archchem_example_input}
\alias{archchem_example_input}
\title{Example input data for ASTR}
\format{
A data frame with 15 observations and 53 variables.
}
\usage{
data(archchem_example_input)
}
\description{
Example input data for ASTR
}
\keyword{datasets}
