% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{ast_cp}
\alias{ast_cp}
\title{A color palette for general plotting and Shiny app use.}
\format{
A character vector of colors
}
\source{
data-raw/create_color_palettes.R
}
\usage{
ast_cp
}
\description{
\itemize{
\item #CC689A - purple1
\item #0099CC - blue1
\item #FF3366 - pink1
\item #33CC99 - green1
\item #FF6633 - orange1
\item #FFCC00 - yellow1
\item #CC0066 - pink2
\item #175E46 - green2
\item #339999 - blue2
\item #FF9900 - gold
\item #330099 - purple2
\item #70AFC4 - blue3
\item #664d99 - purple3
\item #CC6633 - orange2
\item #801832 - burgundy
\item #FFE380 - yellow2
}
}
\details{
\itemize{
\item scale_fill_manual(values = ast_cp)
\item To use for line and point colors, add
scale_color_manual(values=ast_cp)
}
}
\keyword{datasets}
