% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_decompose.R
\name{ast_decompose}
\alias{ast_decompose}
\title{Decompose an R function}
\usage{
ast_decompose(x)
}
\arguments{
\item{x}{a function}
}
\value{
result of \code{\link[utils:getParseData]{utils::getParseData()}}, a data.frame; but
we wrap it in a thin S3 wrapper to make it easy to see how the
function changes as you modify it with \code{\link[=ast_modify]{ast_modify()}}
}
\description{
Decompose an R function
}
\examples{
foo <- function(x) {
  x + 1
}
foo(5)
df <- ast_decompose(foo)
df
}
