% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_recompose.R
\name{ast_recompose}
\alias{ast_recompose}
\title{Recompose an R function}
\usage{
ast_recompose(x, as_expr = FALSE)
}
\arguments{
\item{x}{an object of class \code{ast}}

\item{as_expr}{(logical) return the function as an expression?
default: \code{FALSE}}
}
\value{
a function as a character string, or an expression if
\code{as_expr=TRUE}
}
\description{
Recompose an R function
}
\examples{
foo <- function(x) {
  x + 1
}
foo(5)

# decompose the function
df <- ast_decompose(foo)
df

# recompose the function
fun <- ast_recompose(df)
fun

parse(text = fun)
eval(parse(text = fun))
eval(parse(text = fun))(5)
}
