% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_utils.R
\name{ASTR_to_long}
\alias{ASTR_to_long}
\title{Transforms ASTR format into long format for plotting with ggplot2}
\usage{
ASTR_to_long(df, aesthetic, value)
}
\arguments{
\item{df}{the data frame to be transformed into long format}

\item{aesthetic}{the aesthetic that contains the values from ASTR}

\item{value}{the name of the column with the values after transformation into
long format}
}
\description{
The function expects that a data frame is passed as aesthetic to ggplot,
resulting in a tibble with a nested dataframe with the name of the aesthtic.
}
\keyword{internal}
