% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_conversion_atomic.R
\name{atomic_conversion}
\alias{atomic_conversion}
\alias{wt_to_at}
\alias{at_to_wt}
\title{Conversion between wt\% and at\%}
\usage{
wt_to_at(df, elements = colnames(get_unit_columns(df, "wtP")), drop = TRUE)

at_to_wt(df, elements = colnames(get_unit_columns(df, "atP")), drop = TRUE)
}
\arguments{
\item{df}{Data frame with compositional data.}

\item{elements}{character vector with the chemical symbols of the elements
that should be converted. Default are all columns in an \href{ASTR}{\verb{ASTR object}} in the unit to be converted. See \emph{Details} for further
information}

\item{drop}{If \code{TRUE}, the default, columns with unconverted values are
dropped. If false, columns with unconverted values are kept and a suffix
added to the column names of the converted values.
\itemize{
\item \verb{_at} for conversions to atomic percent
\item \verb{_wt} for conversions to weight percent.
}}
}
\value{
The original data frame with the converted concentrations normalised
to 100\%.
}
\description{
Convert chemical compositions between weight percent (\emph{wt\%}) and atomic
percent (\emph{at\%}). Results are always normalised to 100\%.
}
\details{
The column names of the elements to be converted must be equivalent to their
chemical symbols. The functions convert only values in \emph{wt\%} or \emph{at\%}. If
concentrations are present in another concentration unit (e.g. \emph{ppm},
\emph{µg/kg}), run \code{\link[=unify_concentration_unit]{unify_concentration_unit(df, "wtP")}} first to convert all concentrations to
\emph{wt\%}. If \code{df} is an \href{ASTR}{\verb{ASTR object}}, only elements will be converted
to \emph{at\%}, and oxides in \emph{wt\%} are automatically excluded. To convert oxides
into \emph{at\%} and vice versa, convert to \emph{wt\%} first.
}
\examples{

library(magrittr)

# Convert weight percent to atomic percent and to weight percent
df <- data.frame(ID = 1, Si = 46.74, O = 53.26)  # SiO2 composition
at <- wt_to_at(df, elements = c("Si", "O"))
at_to_wt(at, elements = c("Si", "O"))

# preserve columns with unconverted values
wt_to_at(df, elements = c("Si", "O"), drop = FALSE)

# Use with ASTR objects
# Create ASTR object
test_file <- system.file("extdata", "test_data_input_good.csv", package = "ASTR")
arch <- read_ASTR(test_file, id_column = "Sample", context = 1:7)

# Convert columns in wt\% to at\%
arch_atP <- wt_to_at(arch)

# To convert all applicable concentrations, unify units first:
arch_all <- unify_concentration_unit(arch, "wtP") \%>\%
  wt_to_at()

# Elements already present in the converted unit are ignored.
rowSums(get_unit_columns(arch_all, "atP")[-1], na.rm = TRUE) > 100

}
