% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_bdl_strategies.R
\name{bdl_strategy_negative}
\alias{bdl_strategy_negative}
\title{Replace negative values with a uniform value}
\usage{
bdl_strategy_negative(cols, value = NA_real_)
}
\arguments{
\item{cols}{The names of the columns to be checked according to the ASTR
conventions.}

\item{value}{The replacement value. Default is to \code{NA_real_}.}
}
\value{
A function
}
\description{
This function generates a function for the argument \code{bdl_strategy} in \link{ASTR}
that replaces negative values in specified columns with a uniform value,
usually \code{NA} or \code{0}.
}
