% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_numeric_columns}
\alias{check_numeric_columns}
\title{TODO: improve when specific  archem classes are defined
TODO: evaluate if worth replacing the generic R code columns[!sapply(df\link{columns}, is.numeric)]
with columns <- df \%>\%
select(all_of(columns)) \%>\%
select(where(~ !is.numeric(.))) \%>\%
names()}
\usage{
check_numeric_columns(df, columns)
}
\description{
TODO: improve when specific  archem classes are defined
TODO: evaluate if worth replacing the generic R code columns[!sapply(df\link{columns}, is.numeric)]
with columns <- df \%>\%
select(all_of(columns)) \%>\%
select(where(~ !is.numeric(.))) \%>\%
names()
}
