% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_copper_alloy_classification_pollard.R
\name{copper_alloy_pollard}
\alias{copper_alloy_pollard}
\title{Copper alloy classification according to Pollard et al. (2015).}
\usage{
copper_alloy_pollard(
  df,
  elements = c(Sn = "Sn", Zn = "Zn", Pb = "Pb"),
  id_column = "ID",
  group_as_symbol = FALSE,
  ...
)
}
\arguments{
\item{df}{data frame with the data to be classified.}

\item{elements}{named character vector with column names of Sn, Zn, and Pb.}

\item{id_column}{name of the column in \code{df} with the identifiers of each row.
Default to \code{ID}.}

\item{group_as_symbol}{logical. If \code{FALSE}, the default, copper groups are
reported as their label. Otherwise, copper groups are reported by their
symbol.}

\item{...}{Additional arguments for unit conversion, see \link{atomic_conversion}
and \link{oxide_conversion} for details.}
}
\value{
If \code{df} is an \link[=ASTR]{ASTR object}, the output is an object of the
same type including the ID column, the contextual columns, the elements
used for classification and the alloy type. In all other cases, the data
frame provided as input with the column for the alloy type.
}
\description{
Classification of copper alloy artefacts following the system
proposed in Pollard et al. (2015) based Sn, Zn, and Pb concentrations in
wt\%.
}
\examples{
sample_df <- data.frame(
  ID = 1:8,
  Sn = c(0.5, NA, 5, 5, 0.5, 5, 5, 5),
  Zn = c(0.5, 0.5, 0.5, 0.5, 5, 5, 0.5, 5),
  Pb = c(0.5, 5, 0.5, 5, 0.5, 0.5, 5, 5)
)
copper_alloy_pollard(sample_df)

# For ASTR objects, units and oxides are automatically converted
sample_df <- as_ASTR(
  data.frame(
    ID = 1:8,
    SnO_wtP = c(0.5, 0.5, 5, 5, 0.5, 5, 5, 5),
    ZnO_wtP = c(0.5, 0.5, 0.5, 0.5, 5, 5, 0.5, 5),
    PbO_wtP = c(0.5, 5, 0.5, 5, 0.5, NA, 5, 5)
  )
)
copper_alloy_pollard(sample_df, elements = c(Sn = "SnO", Zn = "ZnO", Pb = "PbO"))

}
\references{
Pollard, A. M., Bray, P., Gosden, C., Wilson, A., and Hamerow, H.
(2015). Characterising copper-based metals in Britain in the first
millennium AD: a preliminary quantification of metal flow and recycling.
Antiquity 89(345), pp. 697-713. \url{https://doi.org/10.15184/aqy.2015.20}
}
\seealso{
Other copper alloy classifications:
\code{\link[=copper_alloy_bb]{copper_alloy_bb()}},
\code{\link[=copper_group_bray]{copper_group_bray()}}
}
\concept{copper alloy classifications}
