% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_header_rows}
\alias{count_header_rows}
\title{Count header rows in an Access Sensor Technologies (AST) air sampler log file}
\usage{
count_header_rows(file)
}
\arguments{
\item{file}{Any AST air sampler log file name}
}
\value{
A data frame including:
\itemize{
\item Header row counts with and without blank rows
\item Row count to the start of diagnostic summary, if applicable
\item Number of rows in diagnostic summary, if applicable
\item Boolean indicating if the file is diagnostic
}
}
\description{
\code{count_header_rows()} is a quick way to find where the log file header ends
before reading the file into a data frame.
It finds the number of rows in the log file header with and without blank
rows included. This information is useful for determining the number of
header rows to read or skip when using file read functions from other
packages such as fread, read_csv, and read.csv. For diagnostic files, this
function indicates the starting row and length of the diagnostic summary too.
\code{count_header_rows()} also returns a boolean value to indicate if the file is
diagnostic.
}
\examples{
# Using a standard UPASv2x log file
upas_filename <- 'PSP00270_LOG_2024-06-10T21_50_55UTC_name____________eng_______.txt'
file <- system.file("extdata", upas_filename, package = "astr", mustWork = TRUE)
count_header_rows(file)

# Using a diagnostic UPASv2x log file
upas_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
file <- system.file("extdata", upas_filename, package = "astr", mustWork = TRUE)
count_header_rows(file)
}
