% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{elements}
\alias{elements}
\title{Elements}
\format{
a vector
}
\description{
List of elements, sorted according to alphabet
}
\seealso{
Other chemical_reference_data: 
\code{\link{isotopes}},
\code{\link{oxides}}
}
\concept{chemical_reference_data}
