% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_conversion_error.R
\name{error_conversion}
\alias{error_conversion}
\alias{rel_to_abs}
\alias{abs_to_rel}
\title{Convert between relative and absolute analytical uncertainties}
\usage{
rel_to_abs(df)

abs_to_rel(df)
}
\arguments{
\item{df}{An ASTR object}
}
\value{
An ASTR object with converted analytical precision columns. The
unchanged input, if it does not contain columns of the respective
analytical precision type
}
\description{
Convert relative to absolute analytical uncertainties and vice versa in ASTR
objects. Work only for objects of class \code{ASTR}.
}
\examples{
test_file <- system.file("extdata", "test_data_input_good.csv", package = "ASTR")
arch <- read_ASTR(test_file, id_column = "Sample", context = 1:7)

arch2 <- abs_to_rel(arch)

arch3 <- rel_to_abs(arch2)

# Conversion is lossless
all.equal(arch$`SiO2_errSD`, arch3$`SiO2_errSD`)

}
