% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_map.R
\name{format_gps_map_data}
\alias{format_gps_map_data}
\title{Formats an Access Sensor Technologies (AST) UPAS v2.1 PLUS log data frame for
generating a gps map}
\usage{
format_gps_map_data(df, variable = c("PM2_5MC", "CO2"))
}
\arguments{
\item{df}{A log data frame returned by the \code{\link[=read_ast_log]{read_ast_log()}} function
with the argument \code{update_names = TRUE}.}

\item{variable}{A column name from the log file. Choose \code{"PM2_5MC"} for PM2.5,
and \code{"CO2"} for CO2 data.}
}
\value{
A modified data frame with 30 second mean data for select variables.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{format_gps_map_data()} formats data from an Access Sensor Technologies
UPAS v2.1 PLUS log data frame for the \code{\link[=gps_map]{gps_map()}} function.
This function calculates 30 second averaged PM2.5 or CO2 data and removes
variables not necessary for generating the map.
If UPASv2 data is provided, data will not be generated for that device.
This function calculates a 30 second average for any data logged
at a less than 30 second log interval.
The UPAS must have been outside with GPS reception and collecting PM2.5 or CO2 data.
}
\examples{
multiple_upas_logs <- system.file("extdata", package = "astr", mustWork = TRUE) |>
    list.files(pattern="^PS.*.txt$", full.names = TRUE) \%>\%
    lapply(read_ast_log, update_names=TRUE) \%>\%
    dplyr::bind_rows()

gps_map_data <- format_gps_map_data(multiple_upas_logs, variable="PM2_5MC")
gps_map_data <- format_gps_map_data(multiple_upas_logs, variable="CO2")
}
