% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhb.R
\name{format_hhb_log}
\alias{format_hhb_log}
\title{Format the sample log data from an Access Sensor Technologies Home Health Box log file}
\usage{
format_hhb_log(log, header, tz = NA, cols_keep = c(), cols_drop = c())
}
\arguments{
\item{log}{A data frame of HHB v2 sample log data returned by the \link{fread_ast_log} function.}

\item{header}{A data frame of HHB v2 header data returned by the \link{read_ast_header} function.}

\item{tz}{Optional: A character string specifying the tz database time zone that should be used to display local times.
See \link{read_ast_log} for additional information.}

\item{cols_keep}{Optional: Provide a character vector specifying the names of a subset of sample log columns to keep.}

\item{cols_drop}{Optional: Provide a character vector specifying the names of a subset of sample log columns to remove.
See \link{read_ast_log} for additional information.}
}
\value{
A data frame of of HHB v2 sample log data that are formatted and ready for analysis.
This data frame will contain one row for each timestamp in the sample log.
}
\description{
Format the sample log data from an Access Sensor Technologies Home Health Box log file
}
