% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhb.R
\name{format_hhb_sample_log}
\alias{format_hhb_sample_log}
\title{Convert sample data from a Home Health Box log file header to long format.}
\usage{
format_hhb_sample_log(log)
}
\arguments{
\item{log}{A HHB v2 sample log data frame returned by \code{\link[=read_ast_log]{read_ast_log()}} or \code{\link[=format_hhb_log]{format_hhb_log()}}}
}
\value{
A data frame with a four rows of HHB v2 sample log data for each row in \code{log}.
}
\description{
\code{format_hhb_sample_log()} Selects the filter and sorbent sample data from a
HHB v2 sample log and converts it to a long format to facilitate evaluation
of sample flow rates over time.
}
\details{
The returned data frame will start with the following columns:

\itemize{
\item HHBserial: A string indicating the serial number of the Home Health Box
\item LogFileName: A string indicating log filename as saved on the microSD card in the HHB
\item SampleName: A string indicating the user-supplied sample name
\item Channel: A string indicating the pump channel ("A", "B", "C", or "D")
\item ChannelType: A string indicating the sample type ("Filter" or "Sorbent")
\item SampleTime: A difftime object indicating the time elapsed since the sample started
\item DateTimeUTC: A POSIXct object indicating the timestamp in Coordinated Universal Time
\item DateTimeLocal: A POSIXct object indicating the timestamp in the user-specified local time zone (if present in \code{log})
}

After the columns listed above, data from all columns in the sample log
that start with "1.", "A.", "B.", "C.", or "D." will appear.  In the returned
data frame, the "1.", "A.", "B.", "C.", or "D." prefix will be removed from
each parameter name. Additionally, some names will be updated as shown in the
table below.

\tabular{ll}{
\strong{Name in sample log}  \tab \strong{Name in returned data frame} \cr
1.BMP390_Press      \tab ExtPress \cr
1.BMP390_Temp       \tab ExtTemp  \cr
D.BMP581_Press      \tab IntPress \cr
D.BMP581_Temp       \tab IntTemp  \cr
D.TotalSorbentVol   \tab TotalCartridgeVol \cr
D.SampledSorbentVol \tab SampledCartridgeVol \cr
C.BMP581_Press      \tab IntPress \cr
C.BMP581_Temp       \tab IntTemp  \cr
C.TotalSorbentVol   \tab TotalCartridgeVol \cr
C.SampledSorbentVol \tab SampledCartridgeVol \cr
A.BMP581Int_Press   \tab IntPress \cr
A.BMP581Int_Temp    \tab IntTemp  \cr
A.BMP581Ext_Press   \tab ExtPress \cr
A.BMP581Ext_Temp    \tab ExtTemp  \cr
B.BMP581Int_Press   \tab IntPress \cr
B.BMP581Int_Temp    \tab IntTemp  \cr
B.BMP581Ext_Press   \tab ExtPress \cr
B.BMP581Ext_Temp    \tab ExtTemp  \cr
}

In the returned data frame, "ExtPress" and "ExtTemp" refer to pressures and
temperatures measured outside of each pump manifold, respectively.
"IntPress" and "IntTemp" refer to pressures and temperatures measured inside
of each pump manifold. All pressures are in units of Pascals and all
temperatures are in units of degrees Celsius.
}
\examples{
hhb_filename <- 'HHB00087_LOG_2025-06-03T20_55UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_log <- read_ast_log(hhb_file)
hhb_sample_logs <- format_hhb_sample_log(hhb_log)
}
