% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhb.R
\name{format_hhb_samples}
\alias{format_hhb_samples}
\title{Convert sample data from a Home Health Box log file header to long format.}
\usage{
format_hhb_samples(header)
}
\arguments{
\item{header}{A HHB v2 header data frame returned by \code{\link[=read_ast_header]{read_ast_header()}} or \code{\link[=format_hhb_header]{format_hhb_header()}}}
}
\value{
A data frame with a four rows of HHB v2 sample data for each row in \code{header}.
}
\description{
\code{format_hhb_samples()} Selects the filter (particulate matter) and sorbent
(gas) sample data from a HHB v2 log file header and converts it to a long
format to facilitate matching of sample times, durations, and volumes with
sample mass or composition data from another data file.
}
\details{
The "HHBslot#" variables from the header data frame will be renamed "PumpPCB"
associated with the appropriate sample channel in the data frame retuned by
this function.

Additionally, the ProgrammedStartTime, ProgrammedRuntime, StartDateTimeUTC,
and EndDateTimeUTC variables associated with each sample channel (A-D) in the
SETUP SUMMARY and SAMPLE SUMMARY sections of the log file header will be
renamed to SampleProgrammedStartTime, SampleProgrammedRuntime,
SampleStartDateTimeUTC, and SampleEndDateTimeUTC in the data frame returned
by this function to distinguish them from the ProgrammedStartTime,
ProgrammedRuntime, StartDateTimeUTC, and EndDateTimeUTC variables associated
with the overall HHB sample.
}
\examples{
hhb_filename <- 'HHB00087_LOG_2025-06-03T20_55UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_header <- read_ast_header(hhb_file)
hhb_samples <- format_hhb_samples(hhb_header)
}
