% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhb.R
\name{format_hhb_sensors}
\alias{format_hhb_sensors}
\title{Convert Alphasense electrochemical sensor data from a Home Health Box sample log to long format.}
\usage{
format_hhb_sensors(log, header, temp = "G.SCD30_Temp")
}
\arguments{
\item{log}{A HHB v2 log data frame returned by \code{\link[=read_ast_log]{read_ast_log()}} or \code{\link[=format_hhb_log]{format_hhb_log()}}}

\item{header}{A HHB v2 header data frame returned by \code{\link[=read_ast_header]{read_ast_header()}} or \code{\link[=format_hhb_header]{format_hhb_header()}}. Each sample log present in \code{log} must also be present in \code{header}.}

\item{temp}{A character string containing the name of the temperature variable to be used to apply any temperature corrections to the Alphasense electrochemical sensor data.
The default is "G.SCD30_Temp".}
}
\value{
A data frame of with one row for each timestamp, Alphasense electrochemical sensor, and algorithm in \code{log}.
}
\description{
\code{format_hhb_sensors()} Selects the Alphasense electrochemical gas sensor data
from a HHB v2 sample log and converts it to a long format.
}
\details{
The data frame returned by this function will include the following columns:
\tabular{llll}{
\strong{Column name} \tab \strong{Class}  \tab \strong{Units} \tab \strong{Description} \cr
HHBserial     \tab character \tab -      \tab Serial identifier for the Home Health Box \cr
LogFileName   \tab character \tab -      \tab Log filename \cr
Position      \tab numeric   \tab -      \tab Position in which the sensor was installed in the gas sensor housing (1, 2, 3, or 4) \cr
ID            \tab character \tab -      \tab Serial identifier for the electrochemical (EC) sensor \cr
Type          \tab character \tab -      \tab Model of the EC sensor \cr
ISB_Gain      \tab numeric   \tab mV/nA  \tab Gain on the Individual Sensor Board (ISB) to which the EC sensor was mounted \cr
Sensitivity   \tab numeric   \tab nA/ppm \tab Working electrode (WE) sensitivity to the pollutant of interest (a calibration constant for the EC sensor) \cr
WEt           \tab numeric   \tab mV     \tab Total WE zero offset (a calibration constant for the EC sensor) \cr
AEt           \tab numeric   \tab mV     \tab Total auxiliary electrode (AE) zero offset (a calibration constant for the EC sensor)  \cr
WEe           \tab numeric   \tab mV     \tab WE electronic offset on the ISB (a calibration constant for the EC sensor) \cr
AEe           \tab numeric   \tab mV     \tab AE electronic offset on the ISB (a calibration constant for the EC sensor) \cr
SampleTime    \tab difftime  \tab s      \tab Time elapsed since the sample started \cr
DateTimeUTC   \tab POSIXct   \tab -      \tab Timestamp in Coordinated Universal Time \cr
G.SCD30_Temp  \tab numeric   \tab C      \tab Temperature measured inside the gas sensor housing by the Sensirion SCD30 sensor \cr
WE            \tab numeric   \tab V      \tab Working electrode voltage \cr
AUX           \tab numeric   \tab V      \tab Auxiliary electrode voltage \cr
Alg           \tab numeric   \tab -      \tab Algorithm used to calculate ppb estimate \cr
ppb           \tab numeric   \tab ppb    \tab Estimated mixing ratio of pollutant estimated using the specified algorithm \cr
}

If the \code{temp} argument was \code{NA} or was specified as either: (a) a column that
did not contain temperature data or (b) a column was not present in \code{log},
the returned data frame will not contain the "G.SCD30_Temp" column nor any
other column of temperature data.

If the \code{temp} argument was specified as something other than "G.SCD30_Temp,"
that column of temperature data will appear in the data frame in the place
of the "G.SCD30_Temp" column listed in the table above.

It is recommended that you specify \code{temp} as a temperature that was measured
inside the gas sensor housing, for example: "G.SCD30_Temp", "G.BMP581_Temp",
or "G.SFA30_Temp".

The temperature variable "SEN55_Temp" is measured inside then Sensirion SEN55
optical particulate matter sensor and, if this sensor was operated in
RH/T/Gas/PMS mode, in which the fan inside the SEN55 draws ambient air
through the SEN55 sensor continuously, the "SEN55_Temp" variable might be the
temperature in \code{log} that most closely reflects the ambient temperature.

The temperature variable "M.BMP581_Temp" is measured on the surface of the
main Home Health Box PCB by a sensor that is vented to the inside of the
sorbent media compartment.

The following temperature variables are measured by sensors that outside the
gas sensor housing and mounted on circuit boards that are enclosed in the
main Home Health Box housing. We recommend using these temperatures for
interpretation of the electrochemical gas sensor data only as a last resort
if none of the other temperature measurements described above are available:
Battery_Temp","M.BMP581_Temp","SEN55_Temp","1.BMP390_Temp","D.BMP581_Temp",
"C.BMP581_Temp","A.BMP581Int_Temp","A.BMP581Ext_Temp","B.BMP581Int_Temp",
"B.BMP581Ext_Temp".
}
\examples{
hhb_filename <- 'HHB00087_LOG_2025-06-03T20_55UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_log <- read_ast_log(hhb_file)
hhb_header <- read_ast_header(hhb_file)
hhb_ec <- format_hhb_sensors(hhb_log, hhb_header)
}
