% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2.R
\name{format_upasv2_log}
\alias{format_upasv2_log}
\title{Read the log data from a UPASv2 log file}
\usage{
format_upasv2_log(df_h, df_raw, tz_offset = NA, update_names = FALSE)
}
\arguments{
\item{df_h}{A UPASv2 header dataframe}

\item{df_raw}{A UPASv2 raw dataframe}

\item{tz_offset}{Pass an option timezone offset.}

\item{update_names}{Option to update old sampler names to latest version.}
}
\value{
A data frame.
}
\description{
Read the log data from a UPASv2 log file
}
\examples{
upasv2_log <- format_upasv2_log(upasv2_header, upasv2_log_raw)
}
