% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_header.R
\name{fread_ast_header}
\alias{fread_ast_header}
\title{Use fread to read the header data from an Access Sensor Technologies air sampler log file}
\usage{
fread_ast_header(file)
}
\arguments{
\item{file}{Any Access Sensor Technologies air sampler log file name.}
}
\value{
A list containing two items named "header" and "diag". The "header"
item is a data table that contains the header data and the "diag" item is a
data table that contains the diagnostic test data. These data tables contain
the data formatted exactly as they appear in the log file.
}
\description{
\code{fread_ast_header} uses fread to read the header data and, if applicable, the
diagnostic test data from the specified file.
}
\examples{
# UPASv2 EXAMPLES
upasv2_filename <- 'PS1771_LOG_2024-06-13T21_20_17UTC_GPSoutside_________Eng.txt'
upasv2_file <- system.file("extdata", upasv2_filename, package = "astr", mustWork = TRUE)
upasv2_header_list <- fread_ast_header(upasv2_file)
upasv2_header <- upasv2_header_list$header

# UPASv2x EXAMPLES
upasv2x_filename <- 'PSP00270_LOG_2024-06-25T21_37_48UTC_GPS-in-out______----------.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_header_list <- fread_ast_header(upasv2x_file)
upasv2x_header <- upasv2x_header_list$header

# Diagnostic file
upasv2x_diag_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
upasv2x_diag_file <- system.file("extdata", upasv2x_diag_filename, package = "astr", mustWork = TRUE)
upasv2x_diag_header_list <- fread_ast_header(upasv2x_diag_file)
upasv2x_diag_header <- upasv2x_diag_header_list$header
upasv2x_diag_diag <- upasv2x_diag_header_list$diag

# HHB EXAMPLES
hhb_filename <- 'HHB00032_LOG_2024-07-01T18_20UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_header <- fread_ast_header(hhb_file)$header
}
