% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_log.R
\name{fread_ast_log}
\alias{fread_ast_log}
\title{Use fread to read the sample log data from an Access Sensor Technologies air sampler log file}
\usage{
fread_ast_log(file)
}
\arguments{
\item{file}{Any Access Sensor Technologies air sampler log file name.}
}
\value{
A data frame of unformatted sample log data
}
\description{
\code{fread_ast_log()} reads in the sample log data exactly as it appears in the file.
}
\examples{
# UPASv2 EXAMPLES
upasv2_filename <- 'PS1771_LOG_2024-06-13T21_20_17UTC_GPSoutside_________Eng.txt'
upasv2_file <- system.file("extdata", upasv2_filename, package = "astr", mustWork = TRUE)
upasv2_log_raw <- fread_ast_log(upasv2_file)

# UPASv2x EXAMPLES
upasv2x_filename <- 'PSP00270_LOG_2024-07-11T18_01_22UTC_PM_CO2_Map______----------.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_log_raw <- fread_ast_log(upasv2x_file)

# Diagnostic file
upasv2x_diag_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
upasv2x_diag_file <- system.file("extdata", upasv2x_diag_filename, package = "astr",
                                  mustWork = TRUE)
upasv2x_diag_log_raw <- fread_ast_log(upasv2x_diag_file)

# HHB EXAMPLES
hhb_filename <- 'HHB00032_LOG_2024-07-01T18_20UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_log_raw <- fread_ast_log(hhb_file)
}
