% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stacey_karmers.R
\name{geom_stacey_kramers}
\alias{geom_stacey_kramers}
\alias{geom_sk_labels}
\title{Stacey-Kramers Lead Evolution Geom}
\usage{
geom_stacey_kramers(
  mapping = NULL,
  data = NULL,
  system = "76",
  Mu1 = 10,
  Kappa = 4,
  ...
)

geom_sk_labels(
  mapping = NULL,
  data = NULL,
  system = "76",
  Mu1 = 10,
  Kappa = 4,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{system}{Character "76" or "86" defining the isotope system.}

\item{Mu1}{Second-stage 238U/204Pb ratio (default 10).}

\item{Kappa}{Second-stage 232Th/238U ratio (default 4).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{Ti}{Initial time of the second stage in years (default 3.7Ga).}

\item{interval}{Time interval for isochron labels in years (default 200Ma).}

\item{show_geochron}{Logical; should the Geochron be plotted?}

\item{show_isochrons}{Logical; should time isochrons be plotted?}

\item{kappa_list}{Numeric vector of Kappa values to plot in "86" system.}
}
\description{
Stacey-Kramers Lead Evolution Geom
}
