% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tz_string}
\alias{get_tz_string}
\title{Get the time zone string to be used to format local times in an Access Sensor Technologies (AST) air sampler log file}
\usage{
get_tz_string(UTCOffset, tz = NA)
}
\arguments{
\item{UTCOffset}{A numeric value from the log file header representing the
offset, in hours, between the UTC time and the local time zone associated
with the sample. In UPAS log files, this variable is named "GPSUTCOffset".
In  Home Health Box v2 log files, this variable is named "UTCOffset".}

\item{tz}{Optional: A character string specifying the tz database time zone
that should be used to display local times. Example tz database time zones
include: "America/New_York", "America/Denver", and "America/Los_Angeles".
For additional information, see: \url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}}
}
\value{
A string containing the name of a time zone or an \code{NA} value.
}
\description{
\code{get_tz_string()} returns a string representing the time zone with which
local times in a log file should be formatted. If the GPSUTCOffset in your
log file is a whole number of hour, \code{get_tz_string()} function will return a
string. If the GPSUTCOffset in your log file is a fraction of an hour, it's
best to specify the optional \code{tz} parameter; otherwise, \code{get_tz_string()}
will return \code{NA} and local times in the log file will be formatted as strings
instead of as \code{POSIXct} objects.
}
