% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{GlobaLID_ASTR}
\alias{GlobaLID_ASTR}
\title{GloabaLID lead isotope database}
\format{
ASTR object with 1 contextual and 3 Pb Isotope variables
\describe{
\item{Political province/regions}{Regional grouping of lead isotope data}
\item{206Pb/204Pb}{Lead Isotope ratio of 206Pb/204Pb}
\item{207Pb/204Pb}{Lead Isotope ratio of 207Pb/204Pb}
\item{208Pb/204Pb}{Lead Isotope ratio of 208Pb/204Pb}
}
}
\source{
\url{https://globalid.dmt-lb.de/}
}
\usage{
GlobaLID_ASTR
}
\description{
ASTR object containing Lead isotope data along with "Political province/regions"
as grouping variable.
}
\keyword{datasets}
