% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{gps_map}
\alias{gps_map}
\title{Calculate 30 second averages for select variables}
\usage{
gps_map(df)
}
\arguments{
\item{df}{Pass a UPAS v2+ log data frame from 'get_30s_mean' function.}
}
\value{
A leaflet map of 30s averages of the selected variable.
}
\description{
Calculate 30 second averages for select variables
}
\examples{
upasv2x_30s_mean <- df_30s_mean(upasv2x_log)
}
