% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oxide_conversion.R
\name{interactive_oxide_select}
\alias{interactive_oxide_select}
\title{Interactive oxide selection}
\usage{
interactive_oxide_select(elements)
}
\value{
named vector with all oxides for the supplied list of elements
}
\description{
Provides a selection
}
\keyword{internal}
