% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isotopes}
\alias{isotopes}
\title{Isotopes}
\format{
a vector
}
\usage{
isotopes
}
\description{
List of naturally occurring isotopes, retrieved from https://www.ciaaw.org/isotopic-abundances.htm,
sorted according to chemical element and isotope number
}
\seealso{
Other chemical_reference_data: 
\code{\link{elements}},
\code{\link{oxides}}
}
\concept{chemical_reference_data}
\keyword{datasets}
