% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{map_PM2_5MC}
\alias{map_PM2_5MC}
\title{Generate a gps map for the time-resolved PM2.5 concentration}
\usage{
map_PM2_5MC(df)
}
\arguments{
\item{df}{Pass a UPAS v2+ log data frame from 'get_30s_mean' function.}
}
\value{
A leaflet map of 30s averages for the time-resolved PM2.5 concentration.
}
\description{
Generate a gps map for the time-resolved PM2.5 concentration
}
