% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_normalise_geochem.R
\name{normalise_geochem}
\alias{normalise_geochem}
\title{Geochemical normalisation of data}
\usage{
normalise_geochem(df, reference = names(references_geochem))
}
\arguments{
\item{df}{A data frame in wide format.}

\item{reference}{Character string with the normalisation. See Details for
further information.}
}
\value{
If \code{df} is an \link[=ASTR]{ASTR object}, the output is an object of the
same type including the ID column, the contextual columns, the
compositional data that was normalised, and the normalised values of the
respective elements. In all other cases, the data frame provided as input
with columns added for the calculated age model parameters.

The used reference composition is indicated in the column names of the
output by the value of \code{reference}.
}
\description{
Normalises selected elemental concentrations to a reference composition such
as chondrite or MORB.
}
\details{
See \link{references_geochem} for the supported reference compositions.
Normalisation to other reference compositions is possible by adding them as
named vectors, see examples. They will not be stored permanently in the
package and must be exported and reloaded (or their definition included in
the script). If you would like to include a new reference composition in
ASTR, please reach out to the package maintainers or create a pull request in
the \href{https://github.com/archaeothommy/ASTR}{package's GitHub repo} with the
values to be included and a literature reference.

The function converts all elements in the data frame for which a reference
composition is available. For \link[=ASTR]{ASTR objects}, unit conversion is
handled by the function. For all other objects, the user must ensure that
values and reference composition have the same unit.
}
\examples{
df <- data.frame(
  Sample = c("A","B"),
  La = c(10,5),
  Ce = c(20,8)
)

normalise_geochem(
  df,
  elements = c("La","Ce"),
  reference = "chondrite"
)

# For ASTR objects, units are automatically converted
test_file <- system.file("extdata", "test_data_input_good.csv", package = "ASTR")
arch <- read_ASTR(test_file, id_column = "Sample", context = 1:7)

arch_norm <- normalise_geochem(arch, "chondrite")

# adding reference composition for 31X 7835.8A of the CHARM Set
references_geochem$"31X 7835.8A" <- units::set_units(
  value = "\%",
  x = c(P = 0.122, Mn = 0.093, Fe = 0.1, Co = 0.313, Ni = 0.158, Cu = 69.93,
  Zn = 24.83, As = 0.143, Ag = 0.463, Cd = 0.087, Sn = 0.516, Sb = 0.115,
  Pb = 3.15, Bi = 0.112
  )
)

references_geochem$"31X 7835.8A"

}
\seealso{
Other data normalisation functions: 
\code{\link{normalise_100}()},
\code{\link{normalise_data}()},
\code{\link{normalise_element}()},
\code{\link{normalise_sample}()}
}
\concept{data normalisation functions}
