% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oxide_conversion.R
\name{normalise_rows}
\alias{normalise_rows}
\title{Normalise rows to 100\%}
\usage{
normalise_rows(df)
}
\arguments{
\item{df}{A matrix or data frame

Normalises values in a vector to 100\%}
}
\description{
Normalise rows to 100\%
}
\keyword{internal}
