% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_conversion_oxide.R
\name{oxide_conversion}
\alias{oxide_conversion}
\alias{element_to_oxide}
\alias{oxide_to_element}
\title{Oxide conversion functions}
\usage{
element_to_oxide(
  df,
  elements = colnames(get_unit_columns(df, "wtP")),
  oxide_preference,
  which_concentrations = c("all", "major", "minor", "no_trace"),
  normalise = FALSE,
  drop = FALSE
)

oxide_to_element(
  df,
  oxides = colnames(get_unit_columns(df, "wtP")),
  normalise = FALSE,
  drop = FALSE
)
}
\arguments{
\item{df}{Data frame with compositional data.}

\item{elements, oxides}{Character vector with the chemical symbols of the
elements or oxides that should be converted.}

\item{oxide_preference}{String that controls which oxide should be used if an
element forms more than one oxide. Allowed values are: \code{reducing},
\code{oxidising}, \code{ask}, or a named vector mapping the specific oxide to its
element. See details for further information.}

\item{which_concentrations}{Character string that determines by concentration
which of the \code{elements} or \code{oxides} are converted. Allowed values are:
\itemize{
\item \code{all} (convert all elements or oxides; the default)
\item \code{major} (convert elements or oxides with concentrations >= 1 wt\%)
\item \code{minor} (convert elements or concentrations between 0.1 and 1 wt\%).
\item \code{no_trace} (convert elements or oxides with concentration >=0.1 wt\%)
}}

\item{normalise}{If \code{TRUE}, converted concentrations will be normalised to
100\%. Default to \code{FALSE}.}

\item{drop}{If \code{FALSE}, the default, columns with unconverted values are
kept. If \code{TRUE}, columns with unconverted values are dropped. Dropping
column could result in loss of information as this will also drop columns
with values excluded from conversion by the parameter
\code{which_concentrations}.}
}
\value{
The original data frame with the converted concentrations
}
\description{
Convert between element and oxide weight percent (oxide\%) compositions using
pre-compiled conversion factors.
}
\details{
If the dataset includes already an element and its respective oxide, the
conversion leaves the values of the respective oxide or element unaffected.
The functions convert only values in \emph{wt\%}. If concentrations are present in
another concentration unit (e.g. \emph{ppm}, \emph{µg/kg}), run
\code{\link[=unify_concentration_unit]{unify_concentration_unit(df, "wtP")}} first to
convert all concentrations to \emph{wt\%}. When the input is an
\href{ASTR}{\verb{ASTR object}}, the functions convert #' only elements or oxides with
the respective other type being automatically excluded, even though the unit
for both is \emph{wt\%}. To convert oxides into \emph{at\%} and vice versa, convert to
\emph{wt\%} first.

In \code{element_to_oxide()}, the parameter \code{oxide_preference} controls the
behaviour of the function if the element forms more than one oxide:
\itemize{
\item \code{oxidising}: Use the oxide with the highest oxidation state of the element
(e.g., \code{Fe2O3})
\item \code{reducing}: Use the oxide with the lowest oxidation state of the element
(e.g., \code{FeO})
\item \code{ask}: The user is asked for each element which oxide should be used.
\item named vector: A named vector mapping the oxides to be used to the
elements (e.g., \code{c(Fe = "FeO", Cu = "Cu2O")})
}

In \code{oxide_to_element()}, conversions from different oxides to the same
element (e.g., Fe\if{html}{\out{<sub>}}2\if{html}{\out{</sub>}}O\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}} and FeO to Fe) result in one
column for the element with the sum of all converted values of the respective
element.

Conversion factors are pre-compiled for a wide range of oxides. consequently,
conversion is restricted to the oxides on this list. If you encounter an
oxide that is currently not included, please reach out to the package
maintainers or create a pull request in the \href{https://github.com/archaeothommy/ASTR}{package's GitHub repo} to add it.
}
\examples{

library(magrittr)

# Example data frame with element weight percents
df <- data.frame(ID = "Sample1", Si = 45, Fe = 50, Cr = 5)

# Select elements by oxide_preference
element_to_oxide(df, elements = c("Si", "Fe", "Cr"), oxide_preference = "oxidising")
element_to_oxide(df, elements = c("Fe", "Cr"), oxide_preference = c(Fe = "FeO", Cr = "Cr2O3"))
\dontrun{
element_to_oxide(df, elements = c("Si", "Fe", "Cr"), oxide_preference = "ask")
}

# Conversions are reversible
oxides <- element_to_oxide(
  df,
  elements = names(df[-1]),
  oxide_preference = "oxidising",
  drop = TRUE
)
elements <- oxide_to_element(
  oxides,
  oxides = names(oxides[-1]),
  drop = TRUE
)
all.equal(df, elements)

# Loss of information by using 'which_concentration' to convert a subset when 'drop = TRUE'
df2 <- data.frame(ID = "feldspar", Na = 8.77, Al = 10.29, Si = 32.13, Ba = 0.3, Sr = 0.05)
element_to_oxide(
  df2,
  elements =  names(df2[-1]),
  which_concentration = "major",
  oxide_preference = "reducing",
  drop = TRUE
)

# Conversion from oxide to element summarises columns converting to the same element
df3 <- data.frame(Fe2O3 = 20, FeO = 20, Cr2O3 = 15, CrO2 = 15, CuO = 20, Cu2O = 20)
oxide_to_element(df3, oxides = names(df3), drop = TRUE)

# Use with ASTR objects
# Create ASTR object
test_file <- system.file("extdata", "test_data_input_good.csv", package = "ASTR")
arch <- read_ASTR(test_file, id_column = "Sample", context = 1:7)

# Convert columns from oxide to wt\%
arch_wtP <- element_to_oxide(arch, oxide_preference = "oxidising")

# To convert all applicable concentrations, unify units first:
arch_all <- unify_concentration_unit(arch, "wtP") \%>\%
  element_to_oxide(oxide_preference = "oxidising")

}
