% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxides}
\alias{oxides}
\title{Oxides}
\format{
a vector
}
\usage{
oxides
}
\description{
List of oxides, retrieved from https://www.wikidoc.org/index.php/Oxide,
sorted according to alphabet
}
\seealso{
Other chemical_reference_data: 
\code{\link{conversion_oxides}},
\code{\link{elements}},
\code{\link{isotopes}}
}
\concept{chemical_reference_data}
\keyword{datasets}
