% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_vector_types.R
\name{percent}
\alias{percent}
\alias{is_percent}
\alias{as_percent}
\title{\code{percent} vector}
\usage{
percent(x = double())

is_percent(x)

as_percent(x)
}
\arguments{
\item{x}{...}
}
\value{
An S3 vector of class \code{astr_percent}.
}
\description{
This creates a double vector that represents percentages.
}
\examples{
percent(c(0.25, 0.5, 0.75))
}
