% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isocron.R
\name{plot_isocron76}
\alias{plot_isocron76}
\alias{plot_isocron86}
\title{Isocrong Plot in Base R}
\usage{
plot_isocron76(mu1 = 10, time = c(0, 3700, 200), color = c("grey30", "grey70"))

plot_isocron86(mu1 = 10, kap = 4, color = c("grey30", "grey70"))
}
\arguments{
\item{mu1}{Mu value of the model curve; Default 10}

\item{time}{Vecotr of start, end periods of isocrons, and intervals in million years; Default c(0, 3700, 200)}

\item{color}{Vector of line and lable colors for highlights and regular lines; Default c('grey30', 'grey70')}

\item{kap}{Kapa Value; Default 4}
}
\value{
lines and ablines for existing plot
}
\description{
Draws a isocron plot for 206Pb/204Pb~207Pb/204Pb and
206Pb/204Pb~208Pb/204Pb plots on Base R plot using PBIso package
}
