% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointcoloud_distribution.R
\name{pointcoloud_distribution}
\alias{pointcoloud_distribution}
\title{Comparing Isotope Samples to Reference Data in 3D Space}
\usage{
pointcoloud_distribution(wod, samples = "samples", ref, groupby = "group")
}
\arguments{
\item{wod}{Working data, with 206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb}

\item{samples}{Heading of a column with Samples names from working data. Default "samples".}

\item{ref}{Reference data with 206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb}

\item{groupby}{Heading of a column with reference group names from reference data. Default "group".}
}
\value{
A \code{list} of three elements:

\item{hull_inclusion}{\code{logical}. A Boolean value indicating the \strong{inclusion}
of the working data (sample points) within the convex \strong{hull} of the reference data.}
\item{centroids}{\code{data.frame}. The coordinates (locations) of the \strong{centroids} #' (mean values) for each defined reference group.}
\item{distances}{\code{matrix}. A distance \strong{matrix} where rows represent the working data samples
and columns represent the reference \strong{centroids}, containing the distance of each sample from each centroid.}
}
\description{
This package compares isotope samples to reference data in 3D space to identify isotopic consistency and the possibility of mixing between sources.
}
