% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ast_header.R
\name{read_ast_header}
\alias{read_ast_header}
\title{Read the header data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
read_ast_header(file)
}
\arguments{
\item{file}{Any AST air sampler  log file name. Test change.}
}
\value{
A data frame with header data in wide format.
}
\description{
Read the header data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
file <- system.file("extdata", filename, package = "astr", mustWork = TRUE)
data_header <- read_ast_header(file)
}
