% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{read_ast_log}
\alias{read_ast_log}
\title{Read the full log data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
read_ast_log(file)
}
\arguments{
\item{file}{Any AST air sampler  log file name.}
}
\value{
A data frame with all log data plus some header data appended.
}
\description{
Read the full log data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
file <- system.file("extdata", filename, package = "astr", mustWork = TRUE)
data_ast_log <- read_ast_log(file)
}
