% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2.R
\name{read_upasv2_header}
\alias{read_upasv2_header}
\title{Read the header data from a UPASv2 log file}
\usage{
read_upasv2_header(file, update_names = FALSE)
}
\arguments{
\item{file}{A UPASv2 log file}

\item{update_names}{Convert old log file column names to match current log
file names.}
}
\value{
A data frame.
}
\description{
Read the header data from a UPASv2 log file
}
\examples{
filename <- 'PS0166_LOG_2021-09-29T17_37_09UTC_test_______________---.txt'
file <- system.file("extdata", filename, package = "astr", mustWork = TRUE)
read_upasv2_header(file, update_names=FALSE)
}
