% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2_x.R
\name{read_upasv2x_header}
\alias{read_upasv2x_header}
\title{Read the header data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
read_upasv2x_header(df)
}
\arguments{
\item{df}{Pass a upasv2x dataframe from read_ast_header function.}
}
\value{
A modified data frame with header data in wide format.
}
\description{
Read the header data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
file <- system.file("extdata", filename, package = "astr", mustWork = TRUE)
read_ast_header(file)
}
