% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{references_geochem}
\alias{references_geochem}
\title{Reference compositions}
\format{
A list including sets of chemical compositions often used as
reference composition for e.g. normalisation. Values are stored as named
vectors of their concentration with assigned units (using
\link[units]{set_units}). To retrieve the list of elements and their
concentrations for e.g. the PM, run \code{references_geochem$PM}.
\describe{
\item{chondrite}{CI chondrite composition in ppm, as defined by Sun &
McDonough (1989)}
\item{PM}{Primitive mantle composition in ppm, as defined by Sun &
McDonough (1989)}
\item{NMORB}{Normal Mid-ocean ridge basalt composition in ppm, as defined
by Sun & McDonough (1989)}
\item{EMORB}{Enhanced Mid-Ocean Ridge Basalt composition in ppm, as
defined by Sun & McDonough (1989)}
\item{OIB}{Ocean Island Basalt composition in ppm, as defined by Sun &
McDonough (1989)}
}
}
\usage{
references_geochem
}
\description{
Reference compositions
}
\references{
Sun, S.-S. & McDonough, W.F. (1989). Chemical and isotopic
systematics of oceanic basalts. Geological Society, London, Special
Publications 42, pp.313-345. Table 1, page 318.
\url{https://doi.org/10.1144/gsl.sp.1989.042.01.19}.
}
\keyword{datasets}
