% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_error_conversion.R
\name{rel_to_abs}
\alias{rel_to_abs}
\title{Convert relative errors to absolute errors in ASTR object}
\usage{
rel_to_abs(df)
}
\arguments{
\item{df}{An ASTR object}
}
\value{
An ASTR object with errors converted to absolute
}
\description{
Converts all relative error columns to absolute errors using their
corresponding measured value columns.
}
