% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_numeric_columns}
\alias{return_numeric_columns}
\title{TODO: improve when specific  archem classes are defined, atm it's just a wrapper
TODO: add tests when it does something meaningful
TODO: evaluate if worth replacing the generic R code columns[!sapply(df\link{columns}, is.numeric)]
with df \%>\% select(where(~ !is.numeric(.))) \%>\% names()}
\usage{
return_numeric_columns(df, columns, all = FALSE)
}
\description{
function is supposed to strip any metadata that may influence the ML algorithms (PCA, k-means,...)
}
