% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_flag}
\alias{shiny_flag}
\title{Add a flag to the shiny header indicating a failed sample.
Can be used to highlight rows in the shinyAST app}
\usage{
shiny_flag(df_h)
}
\arguments{
\item{df_h}{Pass a UPAS v2 or v2+ data frame from 'read_ast_header' function.}
}
\value{
A modified data frame with a flag indicating a failed sample.
}
\description{
Add a flag to the shiny header indicating a failed sample.
Can be used to highlight rows in the shinyAST app
}
\examples{
# upasv2x_header_flagged <- shiny_flag(upasv2x_header)
# upasv2_header_flaggged <- shiny_flag(upasv2_header)
}
