% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_header}
\alias{shiny_header}
\title{Rename UPAS header file data frame columns
to be more user friendly for the Shiny app}
\usage{
shiny_header(df_h, fract_units = FALSE)
}
\arguments{
\item{df_h}{Pass a UPAS v2 or v2+ data frame from \link{read_ast_header} function.}

\item{fract_units}{Boolean to specify if units should be fractional (L min^-1 vs L/min).}
}
\value{
A modified data frame with units and user friendly column names for header data.
}
\description{
Rename UPAS header file data frame columns
to be more user friendly for the Shiny app
}
\examples{
# upasv2x_header_shiny <- shiny_header(upasv2x_header)
# upasv2_header_shiny <- shiny_header(upasv2_header)
}
