% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_success_flag}
\alias{shiny_success_flag}
\title{Add a flag to a UPAS v2 or UPAS v2.1 PLUS header data frame to indicate PASS/FAIL for a sample.}
\usage{
shiny_success_flag(df_h)
}
\arguments{
\item{df_h}{A formatted data frame of UPAS v2 or UPAS v2.1 PLUS header data returned by the \code{\link[=read_ast_header]{read_ast_header()}} function.}
}
\value{
A data frame of UPAS v2 or UPAS v2.1 PLUS header data with an added column to indicate sample PASS/FAIL
based off the sample \code{ShutdownMode}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
multiple_upas_headers <- system.file("extdata", package = "astr", mustWork = TRUE) |>
    list.files(pattern="^PS.*.txt$", full.names = TRUE) \%>\%
    lapply(read_ast_header, update_names = TRUE) \%>\%
    dplyr::bind_rows()

upas_headers_flagged <- shiny_success_flag(multiple_upas_headers)
}
