% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_units}
\alias{shiny_units}
\title{Reformat units
to be more user friendly for the Shiny app}
\usage{
shiny_units(df_h)
}
\arguments{
\item{df_h}{Pass a UPAS v2 or v2+ data frame from 'read_ast_header' function.}
}
\value{
A modified data frame with units and user friendly column names for header data.
}
\description{
Reformat units
to be more user friendly for the Shiny app
}
\examples{
upasv2x_header_shiny <- shiny_header(upasv2x_header)
upasv2_header_shiny <- shiny_header(upasv2_header)
}
