% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{special_oxide_states}
\alias{special_oxide_states}
\title{Special oxide states}
\format{
a vector
}
\usage{
special_oxide_states
}
\description{
List of values that are treated like oxides, but are not chemical oxides.
}
\seealso{
Other chemical reference data:
\code{\link{conversion_oxides}},
\code{\link{elements_data}},
\code{\link{isotopes_data}},
\code{\link{oxides_data}}
}
\concept{chemical reference data}
\keyword{datasets}
