% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{standard_groups}
\alias{standard_groups}
\title{Geochemical and other standard groups of elements, oxides, or isotopes}
\format{
A list of sets of elements, oxides, or isotopes that represent
geochemical groups or other commonly used groups. To retrieve the full list
of e.g. the HFSE elements, run \code{standard_groups$HFSE}.
\describe{
\item{REE}{Rare Earth Elements, taken from Periodic table of elements}
\item{HFSE}{High field-strength elements as listed in Salters (1998)}
\item{LILE}{Large-ion lithophile elements, as listed in Rudnick (1998)}
}
}
\usage{
standard_groups
}
\description{
Geochemical and other standard groups of elements, oxides, or isotopes
}
\references{
Salters, V.J.M. (1998). Elements: High field strength. In:
Geochemistry. Encyclopedia of Earth Science. Springer, Dordrecht.
https://doi.org/10.1007/1-4020-4496-8_101

Rudnick, R.L. (1998). Elements: Large-ion lithophile. In: Geochemistry.
Encyclopedia of Earth Science. Springer, Dordrecht.
https://doi.org/10.1007/1-4020-4496-8_104
}
\keyword{datasets}
