% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_Standard_Sample_Bracketing.R
\name{standard_sample_bracketing}
\alias{standard_sample_bracketing}
\title{Standard Sample Bracketing}
\usage{
standard_sample_bracketing(
  df,
  values = "values",
  id_col = "ID",
  id_std = "Std",
  pos = 1,
  notation = c("ratio", "delta", "epsilon"),
  sd_input = 1,
  weight_std = c(0.5, 0.5)
)
}
\arguments{
\item{df}{data frame including at least a column with the analysis results
from the machine and a column with the labels for the standard
measurements.}

\item{values}{String with the name of the column in \code{df} with the numeric
results. Default is ´values´.}

\item{id_col}{String with the name of the column in \code{df}with the identifiers
of each row.}

\item{id_std}{String identifying the standard used for bracketing. Default is
"Std".}

\item{pos}{Integer giving the line of the first standard measurement that
opens the first bracket.}

\item{notation}{String that describes the calculation the user wants to be performed:
\itemize{
\item ratio calculates the ratio between the sample and the standard mean
\item delta calculates the delta value of the sample per mille.
\item epsilon calculates the epsilon value of the sample per ten thousand.
}}

\item{sd_input}{integer that gives the number the standard deviation will be multiplied for.}

\item{weight_std}{A vector of length 2 with numeric value giving the weight
assigned to the opening and closing standard of a bracket, respectively.
The sum of both weights must be one. The default \code{0.5} gives the mean of
both standards.}
}
\value{
A data frame with the SSB values plus Standard Error (SE)
}
\description{
Performs Standard Sample Bracketing (SSB) correction for isotope
ratio data.
}
\details{
This function corrects for instrumental drift or mass bias during
isotopic analyses using the Standard Sample Bracketing method. The measured
isotope value of each sample is corrected by the linearly interpolated mass
bias drift between the two standard measurements taken immediately before
and after the sample measurement.
}
\examples{
data <- data.frame(
  ID = c("Std", "Sample_A", "Std", "Sample_A", "Std", "Sample_A", "Std"),
  values = c(16.928, 18.641, 16.932, 18.643, 16.935, 18.642, 16.938)
)

result <- ASTR::standard_sample_bracketing(
 df = data,
 pos = 1,
 notation = "delta"
)
}
\references{
Mason, T. F.D., Weiss, D. J., Horstwood, M., Parrish, R. R.,
Russell, S. S., Mullaneb, E., and Colesa, B. J. (2004) High-precision Cu
and Zn isotope analysis by plasma source mass spectrometry. Journal of
Analytical Atomic Spectrometry 19, pp. 209-217.
\url{https://doi.org/10.1039/B306958C}
}
