% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oxide_conversion.R
\name{sum_duplicates}
\alias{sum_duplicates}
\title{Sum columns with same column name}
\usage{
sum_duplicates(values)
}
\description{
The function recognises and includes column names that in fact have the same
name but were made unique by R functions.
}
\keyword{internal}
