% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_endmembers.R
\name{summary.liaendmembers}
\alias{summary.liaendmembers}
\title{Summary of LIA Endmember groups}
\usage{
\method{summary}{liaendmembers}(object, ...)
}
\arguments{
\item{object}{An object of class "liaendmembers"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing:

\item{Counts}{The number of observations in each group}
\item{Tolerance}{Tolerance values used in the calculation.}
\item{Clamp}{Clamping values used in the calculation.}
\item{Data}{A data frame of the lead isotope ratios that were used.}
}
\description{
Summary of Lead Isotope analysis 'liaendmemebr' object
}
\examples{

# Create data structure

col_names <- c("206Pb/204Pb", "207Pb/204Pb", "208Pb/204Pb")
df <- ASTR::ArgentinaDatabase[
ASTR::ArgentinaDatabase$`Mining area` == "Distrito El Aguilar",
col_names]
ref_data <- as.ref_data(ASTR::ArgentinaDatabase, col_names, "Mining area")

# Create object with class liaendmembers
end_members <- endmembers(
        df,
        col_names,
        tolerance = c(0.01, 0.01),
        clamp = c(Inf, Inf)
)
# Print summary of the liaendmembers object
summary(end_members)
# Get Uclidian distance
isoprov_dist(end_members, ref_data, dist_type = "ed")
# Get Mass Fractionnation Corrected distance
isoprov_dist(end_members, ref_data, dist_type = "mfd")
# Train XGBOOST model on Reference data
\dontrun{ml_model <- train_data(ref_data)}
# Get XGBOOST predicted ml_resutls
isoprov_predict(end_members, ml_model)
}
