% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_utils.R
\name{tag_astr_context}
\alias{tag_astr_context}
\title{Retags new columns as context}
\usage{
tag_astr_context(df, cols)
}
\arguments{
\item{df}{\link[=ASTR]{ASTR object}}

\item{cols}{charater vector of column names.}
}
\description{
Retags new columns as context
}
\keyword{internal}
