% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{tel_dor}
\alias{tel_dor}
\title{LIA points of silver hoard from Tell Dor (Israel)}
\format{
ASTR object with 3 Pb isotope variables
\describe{
\item{206Pb/204Pb}{Lead isotope ratio of 206Pb/204Pb.}
\item{207Pb/204Pb}{Lead isotope ratio of 207Pb/204Pb.}
\item{208Pb/204Pb}{Lead isotope ratio of 208Pb/204Pb.}
}
}
\source{
Eshel, T., Erel, Y., Yahalom-Mack, N., Tirosh, O., and Gilboa, A.
(2019). Lead isotopes in silver reveal earliest Phoenician quest for metals
in the west Mediterranean. Proceedings of the National Academy of Sciences
116(13), 6007–6012. \url{https://doi.org/10.1073/pnas.1817951116}
}
\usage{
tel_dor
}
\description{
Lead isotope ratios of a Phoenician silver hoard from Tel Dor.
Reference data set of isotope ratios using \code{\link[=as_ASTR]{as_ASTR()}}
}
\seealso{
Other Pb isotope functions:
\code{\link{age_models}},
\code{\link[=as_pbiso_ref_data]{as_pbiso_ref_data()}},
\code{\link[=geom_sk_lines]{geom_sk_lines()}},
\code{\link[=pb_iso_endmembers]{pb_iso_endmembers()}},
\code{\link[=pb_iso_prov_dist]{pb_iso_prov_dist()}},
\code{\link[=pb_iso_prov_predict]{pb_iso_prov_predict()}},
\code{\link[=pb_iso_train_data]{pb_iso_train_data()}}
}
\concept{Pb isotope functions}
\keyword{datasets}
