% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_header.R
\name{transpose_ast_header}
\alias{transpose_ast_header}
\title{Transpose the header data from an Access Sensor Technologies air sampler log
file and format the transposed header as a data frame.}
\usage{
transpose_ast_header(header, diag = NULL)
}
\arguments{
\item{header}{A data table containing header data read using the
\link{fread_ast_header} function.}

\item{diag}{An optional data table containing diagnostic test data read using
the \link{fread_ast_header} function. If the diag argument is specified, key
metrics from the diagnostic test will be added to the header data frame.}
}
\value{
A data frame of header data.
}
\description{
\code{transpose_ast_header} Takes the data read using the \link{fread_ast_header}
function and transposes those data so each variable in the header is a unique
column in a data frame.
}
\examples{
# UPASv2 EXAMPLES
#upasv2_rev100_filename <- 'PS1422_LOG_2020-06-02T18_26_25UTC_rev100-norm________---.txt'
#upasv2_rev100_file <- system.file("extdata", upasv2_rev100_filename, package = "astr", mustWork = TRUE)
#upasv2_rev100_header_list <- fread_ast_header(upasv2_rev100_file)
#upasv2_rev100_header_wide <- transpose_ast_header(upasv2_rev100_header_list$header)

#upasv2_rev100_diag_filename <- 'PS1422_LOG_2020-06-02T18_29_11UTC_DIAGNOSTIC____________.txt'
#upasv2_rev100_diag_file <- system.file("extdata", upasv2_rev100_diag_filename, package = "astr", mustWork = TRUE)
#upasv2_rev100_diag_header_list <- fread_ast_header(upasv2_rev100_diag_file)
#upasv2_rev100_diag_header_wide <- transpose_ast_header(upasv2_rev100_diag_header_list$header, upasv2_rev100_diag_header_list$diag)

upasv2_rev125_filename <- 'PS0166_LOG_2021-09-29T17_37_09UTC_test_______________---.txt'
upasv2_rev125_file <- system.file("extdata", upasv2_rev125_filename, package = "astr", mustWork = TRUE)
upasv2_rev125_header_list <- fread_ast_header(upasv2_rev125_file)
upasv2_rev125_header_wide <- transpose_ast_header(upasv2_rev125_header_list$header)

upasv2_rev130_diag_filename <- 'PS1786_LOG_2023-03-02T21_45_43UTC_DIAGNOSTIC____________.txt'
upasv2_rev130_diag_file <- system.file("extdata", upasv2_rev130_diag_filename, package = "astr", mustWork = TRUE)
upasv2_rev130_diag_header_list <- fread_ast_header(upasv2_rev130_diag_file)
upasv2_rev130_diag_header_wide <- transpose_ast_header(upasv2_rev130_diag_header_list$header, upasv2_rev130_diag_header_list$diag)

upasv2_rev138_filename <- 'PS1771_LOG_2024-06-13T21_20_17UTC_GPSoutside_________Eng.txt'
upasv2_rev138_file <- system.file("extdata", upasv2_rev138_filename, package = "astr", mustWork = TRUE)
upasv2_rev138_header_list <- fread_ast_header(upasv2_rev138_file)
upasv2_rev138_header_wide <- transpose_ast_header(upasv2_rev138_header_list$header)

# UPASv2x EXAMPLES
upasv2x_rev81_filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
upasv2x_rev81_file <- system.file("extdata", upasv2x_rev81_filename, package = "astr", mustWork = TRUE)
upasv2x_rev81_header_list <- fread_ast_header(upasv2x_rev81_file)
upasv2x_rev81_header_wide <- transpose_ast_header(upasv2x_rev81_header_list$header)

upasv2x_rev117_filename <- 'PSP00030_LOG_2022-05-11T23_24_01UTC_---------------_----------.txt'
upasv2x_rev117_file <- system.file("extdata", upasv2x_rev117_filename, package = "astr", mustWork = TRUE)
upasv2x_rev117_header_list <- fread_ast_header(upasv2x_rev117_file)
upasv2x_rev117_header_wide <- transpose_ast_header(upasv2x_rev117_header_list$header)

upasv2x_rev110_diag_filename <- 'PSP00055_LOG_2022-03-24T18_05_32UTC_DIAGNOSTIC________________.txt'
upasv2x_rev110_diag_file <- system.file("extdata", upasv2x_rev110_diag_filename, package = "astr", mustWork = TRUE)
upasv2x_rev110_diag_header_list <- fread_ast_header(upasv2x_rev110_diag_file)
upasv2x_rev110_diag_header_wide <- transpose_ast_header(upasv2x_rev110_diag_header_list$header, upasv2x_rev110_diag_header_list$diag)

upasv2x_rev157_filename <- 'PSP00270_LOG_2024-06-25T21_37_48UTC_GPS-in-out______----------.txt'
upasv2x_rev157_file <- system.file("extdata", upasv2x_rev157_filename, package = "astr", mustWork = TRUE)
upasv2x_rev157_header_list <- fread_ast_header(upasv2x_rev157_file)
upasv2x_rev157_header_wide <- transpose_ast_header(upasv2x_rev157_header_list$header)

upasv2x_rev158_diag_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
upasv2x_rev158_diag_file <- system.file("extdata", upasv2x_rev158_diag_filename, package = "astr", mustWork = TRUE)
upasv2x_rev158_diag_header_list <- fread_ast_header(upasv2x_rev158_diag_file)
upasv2x_rev158_diag_header_wide <- transpose_ast_header(upasv2x_rev158_diag_header_list$header, upasv2x_rev158_diag_header_list$diag)
}
