% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2x.R
\name{units_upasv2x_log}
\alias{units_upasv2x_log}
\title{Read the log data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
units_upasv2x_log(df)
}
\arguments{
\item{df}{Pass a upasv2x dataframe from read_ast_header function.}
}
\value{
A modified data frame with units added to column names of log data.
}
\description{
Read the log data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
upasv2x_log_units <- units_upasv2x_log(upasv2x_log)
}
