% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archchem_basic.R
\name{archchem}
\alias{archchem}
\alias{as.archchem}
\alias{read_archchem}
\title{\strong{archchem}}
\usage{
as.archchem(x, ...)

read_archchem(path)
}
\arguments{
\item{x}{a data.frame}

\item{...}{further arguments passed to or from other methods}

\item{path}{path to the file that should be read}
}
\description{
A data format for chemical analysis datasets in archaeology
}
