% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archchem_basic.R
\name{archchem}
\alias{archchem}
\alias{as_archchem}
\alias{read_archchem}
\alias{format.archchem}
\alias{print.archchem}
\alias{remove_units}
\title{\strong{archchem}}
\usage{
as_archchem(df, id_column = "ID", context = c(), ...)

read_archchem(
  path,
  id_column = "ID",
  context = c(),
  delim = "\\t",
  na = c("", "n/a", "NA", "N.A.", "N/A", "na", "-", "n.d.")
)

\method{format}{archchem}(x, ...)

\method{print}{archchem}(x, ...)

remove_units(x, ...)
}
\arguments{
\item{df}{a data.frame}

\item{...}{further arguments passed to or from other methods}

\item{path}{path to the file that should be read}

\item{delim}{A character string with the separator for tabular data. Use
\verb{\\t} for tab-separated data. Must be provided for all file
types except \code{.xlsx} or \code{.xls}.}

\item{na}{Character vector of strings to be interpret as missing values.}

\item{x}{an object of class archchem}
}
\description{
A data format for chemical analysis datasets in archaeology
}
