% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archchem_basic.R
\name{archchem}
\alias{archchem}
\alias{as.archchem}
\alias{read_archchem}
\alias{format.archchem}
\alias{print.archchem}
\alias{remove_units}
\title{\strong{archchem}}
\usage{
as.archchem(df, ...)

read_archchem(path)

\method{format}{archchem}(x, ...)

\method{print}{archchem}(x, ...)

remove_units(x, ...)
}
\arguments{
\item{df}{a data.frame}

\item{...}{further arguments passed to or from other methods}

\item{path}{path to the file that should be read}

\item{x}{an object of class archchem}
}
\description{
A data format for chemical analysis datasets in archaeology
}
