% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archchem_basic.R, R/archchem_column_select.R,
%   R/archchem_unit_manipulation.R
\name{archchem}
\alias{archchem}
\alias{as_archchem}
\alias{validate}
\alias{read_archchem}
\alias{format.archchem}
\alias{print.archchem}
\alias{get_contextual_columns}
\alias{get_analytical_columns}
\alias{get_isotope_columns}
\alias{get_element_columns}
\alias{get_ratio_columns}
\alias{get_concentration_columns}
\alias{remove_units}
\alias{unify_concentration_unit}
\title{\strong{archchem}}
\usage{
as_archchem(
  df,
  id_column = "ID",
  context = c(),
  bdl = c("b.d.", "bd", "b.d.l.", "bdl", "<LOD", "<"),
  bdl_strategy = function() {
     NA_character_
 },
  guess_context_type = TRUE,
  na = c("", "n/a", "NA", "N.A.", "N/A", "na", "-", "n.d.", "n.a.", "#DIV/0!", "#VALUE!",
    "#REF!", "#NAME?", "#NUM!", "#N/A", "#NULL!"),
  drop_columns = FALSE,
  validate = TRUE,
  ...
)

validate(x, quiet = TRUE, ...)

read_archchem(
  path,
  id_column = "ID",
  context = c(),
  delim = "\\t",
  guess_context_type = TRUE,
  na = c("", "n/a", "NA", "N.A.", "N/A", "na", "-", "n.d.", "n.a.", "#DIV/0!", "#VALUE!",
    "#REF!", "#NAME?", "#NUM!", "#N/A", "#NULL!"),
  bdl = c("b.d.", "bd", "b.d.l.", "bdl", "<LOD", "<"),
  bdl_strategy = function() {
     NA_character_
 },
  drop_columns = FALSE,
  validate = TRUE
)

\method{format}{archchem}(x, ...)

\method{print}{archchem}(x, ...)

get_contextual_columns(x, ...)

get_analytical_columns(x, ...)

get_isotope_columns(x, ...)

get_element_columns(x, ...)

get_ratio_columns(x, ...)

get_concentration_columns(x, ...)

remove_units(x, ...)

unify_concentration_unit(x, unit, ...)
}
\arguments{
\item{df}{a data.frame containing the input table}

\item{id_column}{name of the ID column. Defaults to "ID"}

\item{context}{columns that provide contextual (non-measurement) information;
may be column names, integer positions, or a logical inclusion vector}

\item{bdl}{strings representing “below detection limit” values. By default,
the following are recognized: "b.d.", "bd", "b.d.l.", "bdl", "<LOD", "<"}

\item{bdl_strategy}{function used to replace BDL strings. Defaults to a static
function returning NA}

\item{guess_context_type}{should appropriate data types for contextual columns
be guessed automatically? Defaults to TRUE}

\item{na}{character vector of strings to be interpret as missing values.
By default, the following are recognized: "", "n/a", "NA", "N.A.", "N/A", "na",
"-", "n.d.", "n.a.", "#DIV/0!", "#VALUE!", "#REF!", "#NAME?", "#NUM!", "#N/A",
"#NULL!"}

\item{drop_columns}{should columns that are neither marked as contextual in
\code{context}, nor automatically identified as analytical from the column name,
be dropped to proceed with the reading? Defaults to FALSE}

\item{validate}{should the post-reading input validation be run, which checks
for additional properties of archchem tables. Defaults to TRUE}

\item{...}{further arguments passed to or from other methods}

\item{x}{an object of class archchem}

\item{quiet}{should warnings be printed? Defaults to TRUE}

\item{path}{file path (including extension) to the file to read}

\item{delim}{a character string with the separator for tabular data. Use
\verb{\\t} for tab-separated data. Must be provided for all file types except
\code{.xlsx} or \code{.xls}}

\item{unit}{string with a unit definition that can be understood by
\link[units]{set_units}, e.g. "\%", "kg", or "m/s^2"}
}
\value{
Returns a data structure \code{archchem} which is a tibble-derived object.
}
\description{
A data format for chemical analysis datasets in archaeology,
containing contextual information, numerical elemental, and isotopic data.
\code{read_archchem} reads data from a file (.csv, .xls, .xlsx) into this
format, \code{as_archchem} turns R data.frames to it.
}
\details{
The data files can be fairly freeform, i.e. no specified elements,
oxides, or isotopic ratios are required and no exact order of these needs to
be adhered to. Analyses can contain as many analytical columns as necessary.

The column that contains the unique samples identifier must be specified using
the \code{ID} argument. If the dataset contains duplicate ids they will be renamed
consecutively using the following convention: \verb{_1},\verb{_2}, ... \verb{_n}.

Metadata contained within the dataset must be marked using the \code{context}
argument. If any column in the dataframe is not specified as context and not
recognised as an analytical column, this will result in an error.

Below detection limit notation (i.e. ‘b.d.’, ‘bd’, ‘b.d.l.’, ‘bdl’, ‘<LOD’,
or ‘<..’) for element and oxide concentrations is specified using the \code{bdl}
argument. One or more notations can be used as is appropriate for the dataset,
and can be notations not included in the list above. The argument
\code{bdl_strategy} is used to specify the value for handling detection limits.
This is to facilitate the different handling needs of the detection limit for
future statistical applications, as opposed to automatically assigning such
values as ‘NA’.

Missing values are allowed anywhere in the data file body, and will be replaced
by \code{NA} automatically.
}
