% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{ArgentinaDatabase}
\alias{ArgentinaDatabase}
\title{Lead isotope data fro Argentina}
\format{
\subsection{\code{ArgentinaDatabase}}{

A dataframe with 112 rows and 49 columns
}
}
\source{
\url{https://globalid.dmt-lb.de/}
}
\usage{
ArgentinaDatabase
}
\description{
Lead isotope data from ore deposits in Argentina prepared for the TerraLID database.
}
\keyword{datasets}
