% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_PbIso_ProvPredict.R
\name{as.ref_data}
\alias{as.ref_data}
\title{Create Reference data object for LIA endmember distance and probability estimate functions.}
\usage{
as.ref_data(x, cols, group, min_groupsize = 5)
}
\arguments{
\item{x}{Reference data frame or tibble.}

\item{cols}{Vector of column names which represent 206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb}

\item{group}{Name of the column containing isotope groups as character string.}

\item{min_groupsize}{Integer value for the minimum number samples in a group. (Default = 5)}
}
\value{
ref.data object with Isotope groupings and Isotope columns of
206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb
with shorter names.
}
\description{
Data preprocessing and cleaning ensure the reliability and accuracy of provenance analysis.
This step results in the removal of all NA values from the selected Isotope tables, and groups.
The step also should exclude the groups which have low sample data (in the case of (Shnyr et al., (2026) it was 5).
}
\seealso{
Other Pb isotope functions:
\code{\link{age_models}},
\code{\link[=geom_sk_lines]{geom_sk_lines()}},
\code{\link[=pb_iso_endmembers]{pb_iso_endmembers()}},
\code{\link[=pb_iso_prov_dist]{pb_iso_prov_dist()}},
\code{\link[=pb_iso_train_data]{pb_iso_train_data()}}
}
\concept{Pb isotope functions}
