% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_PbIso_Prov.R
\name{as_pbiso_ref_data}
\alias{as_pbiso_ref_data}
\alias{as_pbiso_ref_data.ASTR}
\title{Create reference data object for LIA endmember distance and probability
estimation functions. Data preprocessing and cleaning ensure the reliability
and accuracy of provenance analysis. This step results in the removal of all
NA values from the selected isotope tables and groups. The step should also
exclude groups with a low number samples (in the case of Shnyr et al. (2026),
this was 5).}
\usage{
as_pbiso_ref_data(x, ...)

\method{as_pbiso_ref_data}{ASTR}(x, group, min_groupsize = 5L, ...)
}
\arguments{
\item{x}{ASTR object containing 206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb isotope
ratios.}

\item{...}{Additional parameters}

\item{group}{Name of the column containing isotope groups as character
string.}

\item{min_groupsize}{Integer value for the minimum number of samples in a
group, default value is \code{5L}.}
}
\value{
\code{ASTR_Pbiso_ref_data} object with isotope groupings and isotope
columns of 206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb.
}
\description{
Create reference data object for LIA endmember distance and probability
estimation functions. Data preprocessing and cleaning ensure the reliability
and accuracy of provenance analysis. This step results in the removal of all
NA values from the selected isotope tables and groups. The step should also
exclude groups with a low number samples (in the case of Shnyr et al. (2026),
this was 5).
}
\examples{
as_pbiso_ref_data(GlobaLID_ASTR,
                  names(GlobaLID_ASTR)[[2]],
                  min_groupsize = 5)

}
\seealso{
Other Pb isotope functions:
\code{\link{age_models}},
\code{\link[=geom_sk_lines]{geom_sk_lines()}},
\code{\link[=pb_iso_endmembers]{pb_iso_endmembers()}},
\code{\link[=pb_iso_prov_dist]{pb_iso_prov_dist()}},
\code{\link[=pb_iso_prov_predict]{pb_iso_prov_predict()}},
\code{\link[=pb_iso_train_data]{pb_iso_train_data()}},
\code{\link{tel_dor}}
}
\concept{Pb isotope functions}
