% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astr-package.R
\docType{package}
\name{astr-package}
\alias{astr}
\alias{astr-package}
\title{astr}
\description{
Package description here
}
\author{
\strong{Maintainer}: Casey Quinn \email{cquinn@accsensors.com} (\href{https://orcid.org/0000-0002-6802-5250}{ORCID})

Authors:
\itemize{
  \item Jessica Tryner \email{jtryner@accsensors.com} (\href{https://orcid.org/0000-0002-0522-4551}{ORCID})
  \item Gabe Neymark \email{gneymark@accsensors.com} (\href{https://orcid.org/0000-0003-0720-1908}{ORCID})
  \item Ander Wilson \email{ander.wilson@colostate.edu} (\href{https://orcid.org/0000-0003-4774-3883}{ORCID})
}

Other contributors:
\itemize{
  \item Access Sensor Technologies [copyright holder, funder]
  \item National Institutes of Health [funder]
}

}
