% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR-package.R
\docType{package}
\name{ASTR-package}
\alias{ASTR}
\alias{ASTR-package}
\title{ASTR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://example.com}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
