% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR-package.R
\docType{package}
\name{ASTR-package}
\alias{ASTR}
\alias{ASTR-package}
\title{ASTR: Archaeometric Standards and Tools in R}
\description{
ASTR defines and implements a community reporting standard for archaeometric datasets. It also provides easy-to-use functions for common plots, statistical analyses, data processing and transformation workflows in archaeometry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://archaeothommy.github.io/ASTR/index.html}
}

}
\author{
\strong{Maintainer}: Thomas Rose \email{roseth@posteo.com} (\href{https://orcid.org/0000-0002-8186-3566}{ORCID})

Authors:
\itemize{
  \item Andrea Acevedo Mejia (\href{https://orcid.org/0009-0002-7441-1737}{ORCID})
  \item Gilberto Artioli (\href{https://orcid.org/0000-0002-8693-7392}{ORCID})
  \item María Florencia Becerra (\href{https://orcid.org/0000-0001-6302-7452}{ORCID})
  \item Adam Kevin Benfer (\href{https://orcid.org/0009-0004-1253-1068}{ORCID})
  \item Lisa Bellemère
  \item Thomas Edward Birch (\href{https://orcid.org/0000-0002-4568-9767}{ORCID})
  \item Karan Desai (\href{https://orcid.org/0009-0008-8224-8435}{ORCID})
  \item Paolo d'Imporzano
  \item Tzilla Eshel (\href{https://orcid.org/0000-0003-0976-0877}{ORCID})
  \item Valerio Gentile (\href{https://orcid.org/0000-0003-4402-2730}{ORCID})
  \item Sabine Klein (\href{https://orcid.org/0000-0002-3939-4428}{ORCID})
  \item Catherine Klesner (\href{https://orcid.org/0000-0002-2264-9383}{ORCID})
  \item Kathryn Murphy (\href{https://orcid.org/0000-0002-5906-7299}{ORCID})
  \item Ben Marwick (\href{https://orcid.org/0000-0001-7879-4531}{ORCID})
  \item Stephen Merkel (\href{https://orcid.org/0000-0001-8730-6923}{ORCID})
  \item Marco Daniele Pelizzari
  \item Alexandra Rodler (\href{https://orcid.org/0000-0002-4087-7160}{ORCID})
  \item Benjamin Sabatini (\href{https://orcid.org/0000-0002-4199-0253}{ORCID})
  \item Clemens Schmid (\href{https://orcid.org/0000-0003-3448-5715}{ORCID})
  \item Berber van der Meulen-van der Veen (\href{https://orcid.org/0000-0001-5297-0269}{ORCID})
  \item Chen Wang (\href{https://orcid.org/0009-0007-1851-2388}{ORCID})
  \item Katrin Julia Westner (\href{https://orcid.org/0000-0001-5529-1165}{ORCID})
}

}
\keyword{internal}
