% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astr-package.R
\docType{package}
\name{astr-package}
\alias{astr-package}
\alias{astr}
\title{astr}
\description{
Decompose and Recompose Functions
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
