% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astr-package.R
\docType{package}
\name{astr-package}
\alias{astr}
\alias{astr-package}
\title{astr: Decompose and Recompose Functions}
\description{
Decompose and Recompose Functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sckott/astr}
  \item \url{https://sckott.github.io/astr/}
  \item Report bugs at \url{https://github.com/sckott/astr/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{sckott@protonmail.com}

}
\keyword{internal}
