% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_basic.R, R/ASTR_column_select.R,
%   R/ASTR_unit_manipulation.R
\name{ASTR}
\alias{ASTR}
\alias{as_ASTR}
\alias{read_ASTR}
\alias{validate}
\alias{format.ASTR}
\alias{print.ASTR}
\alias{get_contextual_columns}
\alias{get_analytical_columns}
\alias{get_isotope_columns}
\alias{get_element_columns}
\alias{get_ratio_columns}
\alias{get_concentration_columns}
\alias{get_error_columns}
\alias{get_unit_columns}
\alias{remove_units}
\alias{unify_concentration_unit}
\title{ASTR Schema implementation}
\usage{
as_ASTR(
  df,
  id_column = "ID",
  context = c(),
  bdl_strategy = bdl_strategy_default,
  guess_context_type = TRUE,
  na = c("", "n/a", "NA", "N.A.", "N/A", "na", "-", "n.d.", "n.a.", "#DIV/0!", "#VALUE!",
    "#REF!", "#NAME?", "#NUM!", "#N/A", "#NULL!"),
  drop_columns = FALSE,
  validate = TRUE,
  ...
)

read_ASTR(
  path,
  id_column = "ID",
  context = c(),
  delim = ",",
  guess_context_type = TRUE,
  na = c("", "n/a", "NA", "N.A.", "N/A", "na", "-", "n.d.", "n.a.", "#DIV/0!", "#VALUE!",
    "#REF!", "#NAME?", "#NUM!", "#N/A", "#NULL!"),
  bdl_strategy = bdl_strategy_default,
  drop_columns = FALSE,
  validate = TRUE,
  ...
)

validate(x, quiet = TRUE, ...)

\method{format}{ASTR}(x, ...)

\method{print}{ASTR}(x, ...)

get_contextual_columns(x, ...)

get_analytical_columns(x, ...)

get_isotope_columns(x, ...)

get_element_columns(x, ...)

get_ratio_columns(x, ...)

get_concentration_columns(x, ...)

get_error_columns(x, ...)

get_unit_columns(x, units, ...)

remove_units(x, ...)

unify_concentration_unit(x, unit, ...)
}
\arguments{
\item{df}{a data.frame containing the input table}

\item{id_column}{name of the ID column. Defaults to "ID"}

\item{context}{columns that provide contextual (non-measurement) information;
may be column names, integer positions, or a logical inclusion vector}

\item{bdl_strategy}{function used to replace "below detection limit" strings.
See \link{bdl_strategies} for the different available strategies and on how to implement
a custom one.}

\item{guess_context_type}{should appropriate data types for contextual
columns be guessed automatically? Defaults to \code{TRUE}}

\item{na}{character vector of strings to be interpret as missing values. By
default, the following are recognized: "", "n/a", "NA", "N.A.", "N/A",
"na", "-", "n.d.", "n.a.", "#DIV/0!", "#VALUE!", "#REF!", "#NAME?",
"#NUM!", "#N/A", "#NULL!"}

\item{drop_columns}{should columns that are neither marked as contextual in
\code{context}, nor automatically identified as analytical from the column name,
be dropped to proceed with the reading? Defaults to \code{FALSE}}

\item{validate}{should the post-reading input validation be run, which checks
for additional properties of ASTR tables. Defaults to \code{TRUE}}

\item{...}{Additional arguments passed to the respective import functions.
See their documentation for details:
\itemize{
\item \code{\link[readxl:read_excel]{readxl::read_excel()}} for file formats \code{.xlsx} or \code{.xls}
\item \code{\link[readr:read_delim]{readr::read_delim()}} for all other file formats.
}}

\item{path}{path to the file that should be read}

\item{delim}{A character string with the separator for tabular data. Must be
provided for all file types except \code{.xlsx} or \code{.xls}. Default to \verb{,}. Use
\verb{\\t} for tab-separated data.}

\item{x}{an object of class ASTR}

\item{quiet}{should warnings be printed? Defaults to \code{TRUE}}

\item{units}{A character vector with units to be selected.}

\item{unit}{string with a unit definition that can be understood by
\link[units]{set_units}, e.g. "\%", "kg", or "m/s^2"}
}
\value{
Returns an object of class \code{ASTR}, which is a tibble-derived object.
}
\description{
A tabular data format for chemical analysis datasets in
archaeology, including contextual information, numerical elemental, and
isotopic data. Columns are assigned units (using \link[units]{set_units})
and categories (in an attribute \code{ASTR_class}) based on the column name. The
following functions allow to create objects of class \code{ASTR}, and to
interact with them.
\itemize{
\item \strong{as_ASTR}: Transforms an R \code{data.frame} to an object of class
\code{ASTR}.
\item \strong{read_ASTR}: Reads data from a file (.csv, .xls, .xlsx) and
converts it into an object of class \code{ASTR}.
\item \strong{validate}: Performs additional validation on \code{ASTR} and returns
a \code{data.frame} as a workable list of potential issues.
\item \strong{get_..._columns}: Subsets \code{ASTR} tables to columns of a certain
category (or \code{ASTR_class}), e.g. only contextual data columns.
\item \strong{remove_units}: Removes unit vector types from the analytical columns
in an \code{ASTR} table and replaces them with simple numeric columns of type
\code{double}.
\item \strong{unify_concentration_unit}: Unifies the unit of each concentration column,
e.g. to either \% or ppm (or any SI unit) to avoid mixing units in derived analyses.
}
As \code{ASTR} is derived from \code{tibble} it is directly compatible with the data
manipulation tools in the tidyverse.
}
\details{
The input data files can be fairly freeform, i.e. no specified
elements, oxides, or isotopic ratios are required and no exact order of
these needs to be adhered to. Analyses can contain as many analytical
columns as necessary.

The column that contains the unique samples identifier must be specified
using the \code{ID} argument. If the dataset contains duplicate IDs they will be
renamed consecutively using the following convention: \verb{_1},\verb{_2}, ... \verb{_n}.

Metadata contained within the dataset must be marked using the \code{context}
argument. If any column in the dataframe is not specified as context and
not recognised as an analytical column, this will result in an error,
unless \code{drop_columns = TRUE} (then it will result in warnings for the
respective columns).

Missing values are allowed anywhere in the data file body, and will be
replaced by \code{NA} automatically.
}
\examples{
library(magrittr)

# reading an ASTR table directly from a file
test_file <- system.file("extdata", "test_data_input_good.csv", package = "ASTR")
arch <- read_ASTR(test_file, id_column = "Sample", context = 1:7)

# turning a data.frame to an ASTR table
test_df <- readr::read_csv(test_file)
arch <- as_ASTR(test_df, id_column = "Sample", context = 1:7)

# validating an ASTR table
validate(arch)

# extracting subsets of columns
conc <- get_concentration_columns(arch) # see also other get_..._columns functions

# unit-aware arithmetics on ASTR columns thanks to the units package
conc$Sb + conc$Ag # works
\dontrun{conc$Sb + conc$Sn} # fails with: cannot convert µg/ml into ppm

# converting units
conc$Sb <- units::set_units(arch$Sb, "ppb") \%>\%
  magrittr::set_attr("ASTR_class", "ASTR_concentration")

# removing all units from ASTR tables
remove_units(arch)

# applying tidyverse data manipulation on ASTR tables
arch \%>\%
  dplyr::group_by(Site) \%>\%
  dplyr::summarise(mean_Na2O = mean(Na2O))
conc_subset <- conc \%>\%
  dplyr::select(-Sn, -Sb) \%>\%
  dplyr::filter(Na2O > units::set_units(4, "wtP"))

# unify all concentration units
unify_concentration_unit(conc_subset, "ppb")

}
