% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atomic_conversion.R
\name{atomic_conversion}
\alias{atomic_conversion}
\alias{wt_to_at}
\alias{at_to_wt}
\title{Conversion between wt\% and at\%}
\usage{
wt_to_at(df, elements, normalise = FALSE, drop = TRUE)

at_to_wt(df, elements, normalise = FALSE, drop = TRUE)
}
\arguments{
\item{df}{Data frame with compositional data.}

\item{elements}{character vector with the chemical symbols of the elements
that should be converted.}

\item{normalise}{If \code{TRUE}, will normalise converted concentrations to 100\%.
Default to \code{FALSE}.}

\item{drop}{If \code{TRUE}, the default, columns with unconverted values are
dropped. If false, columns with unconverted values are kept and a suffix
added to the column names of the converted values.
\itemize{
\item \verb{_at} for conversions to atomic percent
\item \verb{_wt} for conversions to weight percent.
}}
}
\value{
Data frame with the converted concentrations.
}
\description{
Convert chemical compositions between weight percent (wt\%) and atomic percent
(at\%). The column names of the elements to be converted must be equivalent to
their chemical symbols.
}
\examples{
# Convert weight percent to atomic percent
df <- data.frame(Si = 46.74, O = 53.26)  # SiO2 composition
wt_to_at(df, elements = c("Si", "O"))

}
