% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atomic_conversion.R
\name{atomic_conversion}
\alias{atomic_conversion}
\alias{wt_to_at}
\alias{at_to_wt}
\title{Conversion between wt\% and at\%}
\usage{
wt_to_at(df, elements, drop = TRUE)

at_to_wt(df, elements, drop = TRUE)
}
\arguments{
\item{df}{Data frame with compositional data.}

\item{elements}{character vector with the chemical symbols of the elements
that should be converted.}

\item{drop}{If \code{TRUE}, the default, columns with unconverted values are
dropped. If false, columns with unconverted values are kept and a suffix
added to the column names of the converted values.
\itemize{
\item \verb{_at} for conversions to atomic percent
\item \verb{_wt} for conversions to weight percent.
}}
}
\value{
The original data frame with the converted concentrations normalised
to 100\%.
}
\description{
Convert chemical compositions between weight percent (wt\%) and atomic percent
(at\%).
}
\details{
The column names of the elements to be converted must be equivalent to their
chemical symbols. Results are always normalised to 100\%.
}
\examples{
# Convert weight percent to atomic percent and to weight percent
df <- data.frame(Si = 46.74, O = 53.26)  # SiO2 composition
at <- wt_to_at(df, elements = c("Si", "O"))
at_to_wt(at, elements = c("Si", "O"))

# preserve columns with unconverted values
wt_to_at(df, elements = c("Si", "O"), drop = FALSE)

}
