% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atomic_conversion.R
\name{atomic_conversion}
\alias{atomic_conversion}
\alias{wt_to_at}
\alias{at_to_wt}
\title{Conversion between wt\% and at\%}
\usage{
wt_to_at(df, elements, normalise = FALSE, drop = TRUE)

at_to_wt(df, elements, normalise = FALSE, drop = TRUE)
}
\arguments{
\item{df}{Data frame with compositional data.}

\item{elements}{character vector with the chemical symbols of the elements
that should be converted.}

\item{normalise}{If \code{TRUE}, will normalise converted concentration to 100\%.
Default to \code{FALSE}.}

\item{drop}{If \code{FALSE} keeps columns with unconverted values. Default to
\code{TRUE}.}
}
\value{
Data frame with the converted concentrations. If \code{drop = FALSE}, a
suffix is added to the column names with the converted values:
\itemize{
\item \verb{_at\%} for conversions to atomic percent
\item \verb{_wt\%} for conversions to weight percent.
}
}
\description{
Convert chemical compositions between weight percent (wt\%) and atomic percent
(at\%). The column names of the elements to be converted must be equivalent to
their chemical symbols.
}
\examples{
# example code

}
