% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_utils.R
\name{check_required_packages}
\alias{check_required_packages}
\title{Checks if required packages are installed and prompts for installation if needed}
\usage{
check_required_packages(pkgs, .reader = readline)
}
\arguments{
\item{pkgs}{character vector of package names}
}
\description{
Checks if required packages are installed and prompts for installation if needed
}
\keyword{internal}
