% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{conversion_oxides}
\alias{conversion_oxides}
\title{Conversion factors from oxides to elements}
\format{
A data frame with 151 rows and 8 variables:
\describe{
\item{Element}{The symbol of a chemical element.}
\item{AtomicWeight}{The atomic weight (= molar mass) of the respective element.}
\item{Oxide}{The formula of the chemical element's oxide.}
\item{M}{The number of oxygen atoms in the oxide = the number of moles oxygen per mole oxide.}
\item{OxideWeight}{The molar mass of the oxide.}
\item{ElementToOxide}{The factor used in the conversion from the chemical element to its oxide.}
\item{OxideToElement}{The factor used in the conversion from the oxide to its chemical element.}
\item{OxidationState}{The oxidation state of the cation as a numeric value.}
}
}
\usage{
conversion_oxides
}
\description{
Conversion factors from oxides to elements
}
\seealso{
Other chemical reference data:
\code{\link{elements_data}},
\code{\link{isotopes_data}},
\code{\link{oxides_data}},
\code{\link{special_oxide_states}}
}
\concept{chemical reference data}
\keyword{datasets}
