% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conversion_oxides}
\alias{conversion_oxides}
\title{Conversion factors from oxides to elements}
\format{
A data frame with 106 rows and 7 variables:
\describe{
\item{Element}{The symbol of a chemical element.}
\item{AtomicWeight}{The atomic weight (= molar mass) of the respective element.}
\item{Oxide}{The formula of the chemical element's oxide.}
\item{M}{The number of oxygen atoms in the oxide = the number of moles oxygen per mole oxide.}
\item{OxideWeight}{The molar mass of the oxide.}
\item{element_to_oxide}{The factor used in the conversion from the chemical element to its oxide.}
\item{oxide_to_element}{The factor used in the conversion from the oxide to its chemical element.}
}
}
\usage{
conversion_oxides
}
\description{
Conversion factors from oxides to elements
}
\seealso{
Other chemical_reference_data: 
\code{\link{elements}},
\code{\link{isotopes}},
\code{\link{oxides}}
}
\concept{chemical_reference_data}
\keyword{datasets}
