% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_copper_alloy_classification_bb.R
\name{copper_alloy_bb}
\alias{copper_alloy_bb}
\title{Copper alloy classification according to Bayley & Butcher (2004)}
\usage{
copper_alloy_bb(
  df,
  elements = c(Sn = "Sn", Zn = "Zn", Pb = "Pb"),
  id_sample = "ID"
)
}
\arguments{
\item{df}{data frame with the data to be classified.}

\item{elements}{named character vector with column names of Zn, Sn, and Pb
concentrations.}

\item{id_sample}{name of the column in \code{df} with the identifiers of each
row. Default to \code{ID}.}
}
\value{
Original data frame with added column \code{copper_alloy_bb}.
}
\description{
Classification of copper alloy artefacts according to Bayley &
Butcher (2004) based on Zn, Sn, and Pb concentrations in wt\%.
Concentrations must be given in wt\%. Classification uses specific
thresholds and ratios to define alloy types.
}
\examples{
sample_df <- data.frame(
  ID = 1:5,
  Sn = c(5, 1, 4, 0.5, 2),
  Zn = c(12, 20, 6, 2, 10),
  Pb = c(1, 0.5, 5, 9, 12)
)
copper_alloy_bb(sample_df)

}
\references{
Bayley, J. and Butcher, S. (2004). Roman brooches in Britain: a
technological and typological study based on the Richborough Collection.
London: Society of Antiquaries of London.
\url{https://doi.org/10.26530/20.500.12657/50365}
}
