% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_copper_alloy_classification_pollard.R
\name{copper_alloy_pollard}
\alias{copper_alloy_pollard}
\title{Copper alloy classification according to Pollard et al. (2015).}
\usage{
copper_alloy_pollard(
  df,
  elements = c(Sn = "Sn", Zn = "Zn", Pb = "Pb"),
  id_sample = "ID",
  group_as_symbol = FALSE
)
}
\arguments{
\item{df}{data frame with the data to be classified.}

\item{elements}{named character vector with column names of Sn, Zn, and Pb.}

\item{id_sample}{name of the column in \code{df} with the identifiers of each row.
Default to \code{ID}.}

\item{group_as_symbol}{logical. If \code{FALSE}, the default, copper groups are
reported as their label. Otherwise, copper groups are reported by their
symbol.}
}
\value{
The original data frame with the added column \code{copper_alloy_Pollard}.
}
\description{
Classification of copper alloy artefacts following the system
proposed in Pollard et al. (2015) based Sn, Zn, and Pb concentrations in wt\%.
}
\examples{
sample_df <- data.frame(
  ID = 1:8,
  Sn = c(0.5, 0.5, 5, 5, 0.5, 5, 5, 5),
  Zn = c(0.5, 0.5, 0.5, 0.5, 5, 5, 0.5, 5),
  Pb = c(0.5, 5, 0.5, 5, 0.5, 0.5, 5, 5)
)
copper_alloy_pollard(sample_df)

}
\references{
Pollard, A. M., Bray, P., Gosden, C., Wilson, A., and Hamerow, H.
(2015). Characterising copper-based metals in Britain in the first
millennium AD: a preliminary quantification of metal flow and recycling.
Antiquity 89(345), pp. 697-713. \url{https://doi.org/10.15184/aqy.2015.20}
}
