% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_copper_classification_Bray.R
\name{copper_group_bray}
\alias{copper_group_bray}
\title{Copper classification according to Bray et al. (2015)}
\usage{
copper_group_bray(
  df,
  elements = c(As = "As", Sb = "Sb", Ag = "Ag", Ni = "Ni"),
  threshold = 0.1
)
}
\arguments{
\item{df}{data frame with the data to be classified.}

\item{elements}{Named character vector with the column names of the As, Sb, Ag, and Ni concentrations.}

\item{threshold}{Numeric. Values > threshold are interpreted as "present".
Default is 0.1 (as suggested in Bray et al., 2015).}
}
\value{
The original data frame with the added column \code{copper_group_bray}.
}
\description{
Classification of copper artefacts according to Bray et al. (2015) into
one of 16 trace element compositional groups based on As, Sb, Ag, or
Ni being below or above 0.1 wt\%.
}
\references{
Bray, P., Cuénod, A., Gosden, C., Hommel, P., Liu, P. and
Pollard, A. M. (2015), Form and flow: the ‘karmic cycle’ of copper. Journal
of Archaeological Science 56, pp. 202-209.
\url{https://doi.org/10.1016/j.jas.2014.12.013}.
}
