% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{elements_data}
\alias{elements_data}
\title{Chemical elements}
\format{
a vector
}
\usage{
elements_data
}
\description{
List of chemical elements as their symbols, sorted according to alphabet.
}
\seealso{
Other chemical reference data: 
\code{\link{conversion_oxides}},
\code{\link{isotopes_data}},
\code{\link{oxides_data}},
\code{\link{special_oxide_states}}
}
\concept{chemical reference data}
\keyword{datasets}
