% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elements_data}
\alias{elements_data}
\title{Elements}
\format{
a vector
}
\usage{
elements_data
}
\description{
List of elements, sorted according to alphabet
}
\seealso{
Other chemical_reference_data: 
\code{\link{conversion_oxides}},
\code{\link{isotopes_data}},
\code{\link{oxides_data}},
\code{\link{special_oxide_states}}
}
\concept{chemical_reference_data}
\keyword{datasets}
