% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{format_ast_header}
\alias{format_ast_header}
\title{Format the header data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
format_ast_header(df_h_raw, update_names = FALSE)
}
\arguments{
\item{df_h_raw}{Any AST air sampler unformatted header dataframe.}

\item{update_names}{Option to update old sampler names to latest version.}
}
\value{
A data frame with header data in wide format.
}
\description{
Format the header data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
data_ast_header <- format_ast_header(data_ast_raw, update_names=FALSE)
}
