% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{format_ast_log}
\alias{format_ast_log}
\title{Extract only the log data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
format_ast_log(df_raw)
}
\arguments{
\item{df_raw}{Any AST air sampler unformatted log file dataframe.}
}
\value{
A data frame with all log data.
}
\description{
Extract only the log data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
data_ast_log <- format_ast_log(data_ast_raw)
}
