% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_log.R
\name{format_ast_log}
\alias{format_ast_log}
\title{Format the sample log data from an Access Sensor Technologies air sampler log file}
\usage{
format_ast_log(
  log,
  header,
  update_names = FALSE,
  tz = NA,
  cols_keep = c(),
  cols_drop = c(),
  shiny = FALSE
)
}
\arguments{
\item{log}{An unformatted data frame of sample log data.}

\item{header}{A formatted data frame of header data.}

\item{update_names}{Option to update any deprecated variable names from log files recorded using older firmware versions to the variable names used in the current firmware version.
See \link{read_ast_log} for additional information.}

\item{tz}{Optional: A character string specifying the tz database time zone that should be used to display local times.
See \link{read_ast_log} for additional information.}

\item{cols_keep}{Optional: Provide a character vector specifying the names of a subset of sample log columns to keep.}

\item{cols_drop}{Optional: Provide a character vector specifying the names of a subset of sample log columns to remove.
See \link{read_ast_log} for additional information.}

\item{shiny}{Option to make TRUE if using this function with AST shiny app.}
}
\value{
A data frame with formatted sample log data plus key header data appended.
}
\description{
\code{format_ast_log} Applies device-specific formatting to the columns of a
sample log data frame returned by the \link{fread_ast_log} function. It sets the
proper data type for each variable and adds columns that aid in identifying
unique log files when data from multiple sample logs have been combined into
a single data frame.
}
\examples{
data_ast_log <- format_ast_log(upasv2x_header, data_ast_raw)
}
