% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{format_ast_log}
\alias{format_ast_log}
\title{Extract only the log data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
format_ast_log(
  df_h,
  df_raw,
  tz_offset = NA,
  update_names = FALSE,
  cols_keep = c(),
  cols_drop = c(),
  units = FALSE,
  shiny = FALSE
)
}
\arguments{
\item{df_h}{An AST air sampler formatted log file header dataframe.}

\item{df_raw}{Any AST air sampler unformatted log file dataframe.}

\item{tz_offset}{Pass an option timezone offset}
}
\value{
A data frame with all log data.
}
\description{
Extract only the log data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
data_ast_log <- format_ast_log(upasv2x_header, data_ast_raw)
}
