% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_log.R
\name{format_ast_log}
\alias{format_ast_log}
\title{Format the sample log data from an Access Sensor Technologies air sampler log file}
\usage{
format_ast_log(
  log,
  header,
  update_names = FALSE,
  tz = NA,
  cols_keep = c(),
  cols_drop = c()
)
}
\arguments{
\item{log}{An unformatted data frame of sample log data returned by the \code{\link[=fread_ast_log]{fread_ast_log()}} function.}

\item{header}{A formatted data frame of header data returned by the \code{\link[=read_ast_header]{read_ast_header()}} function.}

\item{update_names}{Option to update any deprecated variable names from log
files written using older firmware versions to the variable names used in the
current firmware version. Variable names cannot be updated for files
written using UPAS v2 firmware versions preceding rev100.}

\item{tz}{Optional: A character string specifying the tz database time zone
that should be used to display local times. Example tz database time zones
include: "America/New_York", "America/Denver", and "America/Los_Angeles".
For additional information, see: \url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}

If the GPSUTCOffset in your log file is a fraction of an hour, it's
best to specify the optional \code{tz} parameter; otherwise, local times in the log file
will be incorrectly labeled as UTC time.}

\item{cols_keep}{Optional: A character vector specifying the names of a subset of sample log columns to keep.}

\item{cols_drop}{Optional: A character vector specifying the names of a subset of sample log columns to remove.}
}
\value{
A data frame with formatted sample log data plus key header data appended.
}
\description{
\code{format_ast_log()} Applies device-specific formatting to the columns of a
sample log data frame returned by the \code{\link[=fread_ast_log]{fread_ast_log()}} function. It sets the
proper data type for each variable and adds columns that aid in identifying
unique log files when data from multiple sample logs have been combined into
a single data frame.
}
\examples{
# UPASv2 EXAMPLES
upasv2_filename <- 'PS1771_LOG_2024-06-13T21_20_17UTC_GPSoutside_________Eng.txt'
upasv2_file <- system.file("extdata", upasv2_filename, package = "astr", mustWork = TRUE)
upasv2_log_raw <- fread_ast_log(upasv2_file)
upasv2_header <- read_ast_header(upasv2_file, update_names=FALSE)
upasv2_log <- format_ast_log(upasv2_log_raw, upasv2_header)

# UPASv2x EXAMPLES
upasv2x_filename <- 'PSP00270_LOG_2024-07-11T18_01_22UTC_PM_CO2_Map______----------.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_log_raw <- fread_ast_log(upasv2x_file)
upasv2x_header <- read_ast_header(upasv2x_file, update_names=FALSE)
upasv2x_log <- format_ast_log(upasv2x_log_raw, upasv2x_header, update_names=FALSE)

# Diagnostic file
upasv2x_diag_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
upasv2x_diag_file <- system.file("extdata", upasv2x_diag_filename, package = "astr",
                                  mustWork = TRUE)
upasv2x_diag_log_raw <- fread_ast_log(upasv2x_diag_file)
upasv2x_diag_header <- read_ast_header(upasv2x_diag_file, update_names=FALSE)
upasv2x_diag_log <- format_ast_log(upasv2x_diag_log_raw, upasv2x_diag_header, update_names=FALSE)

# HHB EXAMPLES
hhb_filename <- 'HHB00032_LOG_2024-07-01T18_20UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_log_raw <- fread_ast_log(hhb_file)
hhb_header <- read_ast_header(hhb_file)
hhb_log <- format_ast_log(hhb_log_raw, hhb_header)
}
