% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhb.R
\name{format_hhb_header}
\alias{format_hhb_header}
\title{Format the header data from an Access Sensor Technologies Home Health Box log file}
\usage{
format_hhb_header(df)
}
\arguments{
\item{df}{A HHB v2 header data frame returned by \code{\link[=transpose_ast_header]{transpose_ast_header()}}}
}
\value{
A data frame with a single row of HHB v2 header data that are formatted and ready for analysis.
}
\description{
\code{format_hhb_header()} formats the header data from a HHB v2 log file by setting
the proper data type for each variable.
}
\examples{
hhb_filename <- 'HHB00032_LOG_2024-07-01T18_20UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_header_raw <- fread_ast_header(hhb_file)$header
hhb_header_wide <- transpose_ast_header(hhb_header_raw)
hhb_header <- format_hhb_header(hhb_header_wide)
}
