% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhb.R
\name{format_hhb_log}
\alias{format_hhb_log}
\title{Format the sample log data from an Access Sensor Technologies Home Health Box log file}
\usage{
format_hhb_log(log, header, tz = NA, cols_keep = c(), cols_drop = c())
}
\arguments{
\item{log}{A data frame of HHB v2 sample log data returned by the \code{\link[=fread_ast_log]{fread_ast_log()}} function.}

\item{header}{A data frame of HHB v2 header data returned by the \code{\link[=read_ast_header]{read_ast_header()}} function.}

\item{tz}{Optional: A character string specifying the tz database time zone that should be used to display local times.
See \link{read_ast_log} for additional information.}

\item{cols_keep}{Optional: Provide a character vector specifying the names of a subset of sample log columns to keep.}

\item{cols_drop}{Optional: Provide a character vector specifying the names of a subset of sample log columns to remove.
See \link{read_ast_log} for additional information.}
}
\value{
A data frame of of HHB v2 sample log data that are formatted and ready for analysis.
This data frame will contain one row for each timestamp in the sample log.
}
\description{
Format the sample log data from an Access Sensor Technologies Home Health Box log file
}
\examples{
hhb_filename <- 'HHB00087_LOG_2025-06-03T20_55UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_log_raw <- fread_ast_log(hhb_file)
hhb_header <- read_ast_header(hhb_file)
hhb_log <- format_hhb_log(hhb_log_raw, hhb_header)

# Use of cols_drop, cols_keep, and tz  with a UPASv2x log file
hhb_log_colsdrop <- format_hhb_log(hhb_log_raw, hhb_header,
                cols_drop = c("AccelX", "AccelY", "AccelZ"))
hhb_log_colskeep <- format_hhb_log(hhb_log_raw, hhb_header, tz="America/New_York",
                cols_keep = c("SampleTime", "DateTimeUTC", "DateTimeLocal",
                              "LocalTZ",  "UserTZ",
                              grep("^SEN55", colnames(hhb_log), value=TRUE)))
}
