% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2.R
\name{format_upasv2_header}
\alias{format_upasv2_header}
\title{Read the header data from a UPASv2 log file}
\usage{
format_upasv2_header(df_h, update_names = FALSE)
}
\arguments{
\item{df_h}{A UPASv2 header dataframe}

\item{update_names}{Convert old log file column names to match current log
file names.}
}
\value{
A data frame.
}
\description{
Read the header data from a UPASv2 log file
}
\examples{
upasv2_header <- format_upasv2_header(upasv2_header_raw, update_names=FALSE)
}
