% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2.R
\name{format_upasv2_header}
\alias{format_upasv2_header}
\title{Format UPAS v2 header data that hae already been transposed to a wide data frame}
\usage{
format_upasv2_header(data, update_names = FALSE)
}
\arguments{
\item{data}{A UPASv2 header data frame returned by \link{transpose_ast_header}}

\item{update_names}{Option to update any deprecated variable names from log files recorded using older firmware versions to the variable names used in the current firmware version.

For samples collected using UPAS v2 firmware versions beyond rev100, the old
names shown on the left will be updated to the names shown on the right:
\itemize{
\item VolumetricFlowRate        -> FlowRateSetpoint
\item DutyCycle                 -> FlowDutyCycle
\item LoggedRuntime             -> OverallDuration
\item SampledRuntime            -> PumpingDuration
\item AverageVolumetricFlowRate -> PumpingFlowRateAverage
}

For samples collected using UPAS v2 firmware rev100, the old names shown on
the left will be updated to the names shown on the right:
\itemize{
\item CumulativeSamplingTime -> LifetimeSampleRuntime
\item StartDateTime          -> StartDateTimeUTC
\item AverageVolumetricFlow  -> PumpingFlowRateAverage
}

Variable names cannot be updated for log files written using UPAS v2 firmware
versions preceding rev100.}
}
\value{
A data frame with a single row of UPAS v2 header data that are formatted and ready for analysis.
}
\description{
\code{format_upasv2_header} formats the header data from a UPAS v2 log file.
This function sets the proper data type for each variable, adds a column
specifying the AST sampler type, adds a column describing the shutdown reason
associated with the shutdown mode code, and can be directed to update old
variable names to the current names.
}
\examples{
upasv2_rev100_filename <- 'PS1422_LOG_2020-06-02T18_26_25UTC_rev100-norm________---.txt'
upasv2_rev100_file <- system.file("extdata", upasv2_rev100_filename, package = "astr", mustWork = TRUE)
upasv2_rev100_header_list <- fread_ast_header(upasv2_rev100_file)
upasv2_rev100_header_wide <- transpose_ast_header(upasv2_rev100_header_list$header)
upasv2_rev100_header <- format_upasv2_header(upasv2_rev100_header_wide)

upasv2_rev100_diag_filename <- 'PS1422_LOG_2020-06-02T18_29_11UTC_DIAGNOSTIC____________.txt'
upasv2_rev100_diag_file <- system.file("extdata", upasv2_rev100_diag_filename, package = "astr", mustWork = TRUE)
upasv2_rev100_diag_header_list <- fread_ast_header(upasv2_rev100_diag_file)
upasv2_rev100_diag_header_wide <- transpose_ast_header(upasv2_rev100_diag_header_list$header, upasv2_rev100_diag_header_list$diag)
upasv2_rev100_diag_header <- format_upasv2_header(upasv2_rev100_diag_header_wide)

upasv2_rev125_filename <- 'PS0166_LOG_2021-09-29T17_37_09UTC_test_______________---.txt'
upasv2_rev125_file <- system.file("extdata", upasv2_rev125_filename, package = "astr", mustWork = TRUE)
upasv2_rev125_header_list <- fread_ast_header(upasv2_rev125_file)
upasv2_rev125_header_wide <- transpose_ast_header(upasv2_rev125_header_list$header)
upasv2_rev125_header <- format_upasv2_header(upasv2_rev125_header_wide)

upasv2_rev130_diag_filename <- 'PS1786_LOG_2023-03-02T21_45_43UTC_DIAGNOSTIC____________.txt'
upasv2_rev130_diag_file <- system.file("extdata", upasv2_rev130_diag_filename, package = "astr", mustWork = TRUE)
upasv2_rev130_diag_header_list <- fread_ast_header(upasv2_rev130_diag_file)
upasv2_rev130_diag_header_wide <- transpose_ast_header(upasv2_rev130_diag_header_list$header, upasv2_rev130_diag_header_list$diag)
upasv2_rev130_diag_header <- format_upasv2_header(upasv2_rev130_diag_header_wide)

upasv2_rev138_filename <- 'PS1771_LOG_2024-06-13T21_20_17UTC_GPSoutside_________Eng.txt'
upasv2_rev138_file <- system.file("extdata", upasv2_rev138_filename, package = "astr", mustWork = TRUE)
upasv2_rev138_header_list <- fread_ast_header(upasv2_rev138_file)
upasv2_rev138_header_wide <- transpose_ast_header(upasv2_rev138_header_list$header)
upasv2_rev138_header <- format_upasv2_header(upasv2_rev138_header_wide)
}
