% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2.R
\name{format_upasv2_log}
\alias{format_upasv2_log}
\title{Read the log data from a UPASv2 log file}
\usage{
format_upasv2_log(
  df_h,
  df_raw,
  tz_offset = NA,
  update_names = FALSE,
  units = FALSE
)
}
\arguments{
\item{update_names}{Convert old log file column names to match current log
file names.}

\item{df}{A UPASv2 dataframe}
}
\value{
A data frame.
}
\description{
Read the log data from a UPASv2 log file
}
\examples{
upasv2_log <- format_upasv2_log(upasv2_header, upasv2_log_raw)
}
