% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2x.R
\name{format_upasv2x_header}
\alias{format_upasv2x_header}
\title{Format UPAS v2.1 PLUS header data that have already been transposed to a wide data frame}
\usage{
format_upasv2x_header(data, update_names = FALSE, tz = NA)
}
\arguments{
\item{data}{A UPASv2.1 PLUS header data frame returned by the \code{\link[=transpose_ast_header]{transpose_ast_header()}} function}

\item{update_names}{Option to update any deprecated variable names from log
files written using older firmware versions to the variable names used in the
current firmware version. Variable names cannot be updated for files
written using UPAS v2 firmware versions preceding rev100.}

\item{tz}{Optional: A character string specifying the tz database time zone
that should be used to display local times. Example tz database time zones
include: "America/New_York", "America/Denver", and "America/Los_Angeles".
For additional information, see: \url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}

If the \code{tz} parameter is not specified, local times will be displayed using
the time zone associated with the GPSUTCOffset in the log file.

Note that, in the unlikely event that your local time zone is
"Australia/Lord_Howe," you will need to specify this time zone using the tz
parameter to ensure that your local times are displayed correctly.}
}
\value{
A data frame with a single row of UPAS v2.1 PLUS header data that are formatted and ready for analysis.
}
\description{
\code{format_upasv2x_header()} formats the UPAS v2.1 PLUS log file header data prior
to analysis. This function sets the proper data types for each variable, adds
a column to specify the sampler type, and adds a column to describe the
shutdown reason associated with the shutdown mode code.
}
\examples{
upasv2x_filename <- 'PSP00270_LOG_2024-07-11T18_01_22UTC_PM_CO2_Map______----------.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_header_list <- fread_ast_header(upasv2x_file)
upasv2x_header_wide <- transpose_ast_header(upasv2x_header_list$header)
upasv2x_header <- format_upasv2x_header(upasv2x_header_wide)

# Diagnostic file
upasv2x_diag_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
upasv2x_diag_file <- system.file("extdata", upasv2x_diag_filename, package = "astr",
                                  mustWork = TRUE)
upasv2x_diag_header_list <- fread_ast_header(upasv2x_diag_file)
upasv2x_diag_header_wide <- transpose_ast_header(upasv2x_diag_header_list$header,
                                  upasv2x_diag_header_list$diag)
upasv2x_diag_header <- format_upasv2x_header(upasv2x_diag_header_wide)
}
