% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2x.R
\name{format_upasv2x_log}
\alias{format_upasv2x_log}
\title{Format the sample log data from an Access Sensor Technologies UPAS v2.1 PLUS}
\usage{
format_upasv2x_log(
  log,
  header,
  update_names = FALSE,
  tz = NA,
  cols_keep = c(),
  cols_drop = c()
)
}
\arguments{
\item{log}{A data frame of UPAS v2.1 PLUS sample log data returned by the \link{fread_ast_log} function.}

\item{header}{A data frame of UPAS v2.1 PLUS header data returned by the \link{read_ast_header} function.}

\item{update_names}{Option to update any deprecated variable names from log files recorded using older firmware versions to the variable names used in the current firmware version.

For samples collected using UPAS v2.1 PLUS firmware versions preceding revXXX,
the column name "AceelComplianceHrs" will be updated to "AccelComplianceHrs".

For samples collected using the UPAS v2, the old names shown on the left will
be updated to the names shown on the right:
VolumetricFlowRate -> PumpingFlowRate
\itemize{
\item AtmoRho          -> AtmoDensity
\item FdPdP            -> FilterDP
\item PumpT            -> AtmoT
\item PumpRH           -> AtmoRH
\item PCBT             -> PCB1T
\item PumpP            -> PCB2P
\item PCBP             -> AtmoP
\item GPShdop          -> GPShDOP
\item BFGvolt          -> BattVolt
}

For samples collected specifically using UPAS v2 firmware rev100, the old
names shown on the left will be updated to the names shown on the right:
\itemize{
\item UTCDateTime     -> DateTimeUTC
\item VolFlow         -> VolumetricFlowRate
\item UPASLogFilename -> LogFilename
}

Variable names cannot be updated for log files written using UPAS v2 firmware
versions preceding rev100.}

\item{tz}{Optional: A character string specifying the tz database time zone that should be used to display local times.

Example tz database time zones include: "America/New_York", "America/Denver", and "America/Los_Angeles".
For additional information, see: https://en.wikipedia.org/wiki/List_of_tz_database_time_zones

If your GPSUTCOffset is a whole number of hours, this function will be able
to automatically convert the DateTimeLocal variable in the sample log to a
POSIXct object displayed in the local time zone, without this tz parameter
being specified.  If your GPSUTCOffset is not a whole number of hours, it's
better to specify the local time zone as a character string here.}

\item{cols_keep}{Optional: Provide a character vector specifying the names of a subset of sample log columns to keep.}

\item{cols_drop}{Optional: Provide a character vector specifying the names of a subset of sample log columns to remove.

Column selection will occur in the same order in which the function arguments are specified above.
In other words, the columns specified in cols_keep will be selected first.
If the cols_keep argument is not specified, all columns will be kept.  Then,
The columns specified in cols_drop will be dropped.  If the cols_drop
argument is not specified, no columns will be dropped.}
}
\value{
A data frame of of UPAS v2.1 PLUS sample log data that are formatted and ready for analysis.
This data frame will contain one row for each timestamp in the sample log.
}
\description{
#' @description
\code{format_upasv2x_log} Applies device-specific formatting to the columns of a
sample log data frame returned by the \link{fread_ast_log} function. It sets the
proper data type for each variable and adds columns that aid in identifying
unique log files when data from multiple sample logs have been combined into
a single data frame.
}
\examples{
upasv2x_filename <- 'PSP00270_LOG_2024-06-25T21_37_48UTC_GPS-in-out______----------.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_log_raw <- fread_ast_log(upasv2x_file)
upasv2x_header <- read_ast_header(upasv2x_file, update_names=FALSE)
upasv2x_log <- format_upasv2x_log(upasv2x_log_raw, upasv2x_header, update_names=FALSE)

# Use of cols_drop, cols_keep, and tz  with a UPASv2x log file
upasv2x_log_colsdrop <- format_upasv2x_log(upasv2x_log_raw, upasv2x_header,
                cols_drop = c("DateTimeLocal", "AtmoT", "AtmoP", "AtmoRH"))
upasv2x_log_colskeep <- format_upasv2x_log(upasv2x_log_raw, upasv2x_header, tz="America/New_York",
                cols_keep = c("SampleTime", "DateTimeUTC", "DateTimeLocal",
                               "LocalTZ",  "UserTZ", "AtmoT", "AtmoP", "AtmoRH"))

# Diagnostic file
upasv2x_diag_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
upasv2x_diag_file <- system.file("extdata", upasv2x_diag_filename, package = "astr",
                                  mustWork = TRUE)
upasv2x_diag_log_raw <- fread_ast_log(upasv2x_diag_file)
upasv2x_diag_header <- read_ast_header(upasv2x_diag_file, update_names=FALSE)
upasv2x_diag_log <- format_upasv2x_log(upasv2x_diag_log_raw, upasv2x_diag_header,
                                  update_names=FALSE)
}
