% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_log.R
\name{fread_ast_log}
\alias{fread_ast_log}
\title{Use fread to read the sample log data from an Access Sensor Technologies air sampler log file}
\usage{
fread_ast_log(file)
}
\arguments{
\item{file}{Any Access Sensor Technologies air sampler log file name.}
}
\value{
A data frame of unformatted sample log data
}
\description{
\code{fread_ast_log} reads in the sample log data exactly as it appears in the file.
}
\examples{
upasv2_rev125_filename <- 'PS0166_LOG_2021-09-29T17_37_09UTC_test_______________---.txt'
upasv2_rev125_file <- system.file("extdata", upasv2_rev125_filename, package = "astr", mustWork = TRUE)
upasv2_rev125_log_raw <- fread_ast_log(upasv2_rev125_file)
upasv2_rev130_diag_filename <- 'PS1786_LOG_2023-03-02T21_45_43UTC_DIAGNOSTIC____________.txt'
upasv2_rev130_diag_file <- system.file("extdata", upasv2_rev130_diag_filename, package = "astr", mustWork = TRUE)
upasv2_rev130_diag_log_raw <- fread_ast_log(upasv2_rev130_diag_file)
upasv2_rev138_filename <- 'PS1771_LOG_2024-06-13T21_20_17UTC_GPSoutside_________Eng.txt'
upasv2_rev138_file <- system.file("extdata", upasv2_rev138_filename, package = "astr", mustWork = TRUE)
upasv2_rev138_log_raw <- fread_ast_log(upasv2_rev138_file)
upasv2x_rev81_filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
upasv2x_rev81_file <- system.file("extdata", upasv2x_rev81_filename, package = "astr", mustWork = TRUE)
upasv2x_rev81_log_raw <- fread_ast_log(upasv2x_rev81_file)
upasv2x_rev117_filename <- 'PSP00030_LOG_2022-05-11T23_24_01UTC_---------------_----------.txt'
upasv2x_rev117_file <- system.file("extdata", upasv2x_rev117_filename, package = "astr", mustWork = TRUE)
upasv2x_rev117_log_raw <- fread_ast_log(upasv2x_rev117_file)
upasv2x_rev110_diag_filename <- 'PSP00055_LOG_2022-03-24T18_05_32UTC_DIAGNOSTIC________________.txt'
upasv2x_rev110_diag_file <- system.file("extdata", upasv2x_rev110_diag_filename, package = "astr", mustWork = TRUE)
upasv2x_rev110_diag_log_raw <- fread_ast_log(upasv2x_rev110_diag_file)
upasv2x_rev157_filename <- 'PSP00270_LOG_2024-06-25T21_37_48UTC_GPS-in-out______----------.txt'
upasv2x_rev157_file <- system.file("extdata", upasv2x_rev157_filename, package = "astr", mustWork = TRUE)
upasv2x_rev157_log_raw <- fread_ast_log(upasv2x_rev157_file)
upasv2x_rev158_diag_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
upasv2x_rev158_diag_file <- system.file("extdata", upasv2x_rev158_diag_filename, package = "astr", mustWork = TRUE)
upasv2x_rev158_diag_log_raw <- fread_ast_log(upasv2x_rev158_diag_file)
}
