% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_geom_stacey_karmers.R
\name{geom_sk_lines}
\alias{geom_sk_lines}
\alias{geom_sk_labels}
\title{Stacey-Kramers Lead Evolution Geom}
\usage{
geom_sk_lines(
  mapping = NULL,
  data = NULL,
  system = c("76", "86"),
  Mu1 = 10,
  Kappa = 4,
  ...
)

geom_sk_labels(
  mapping = NULL,
  data = NULL,
  system = "76",
  Mu1 = 10,
  Kappa = 4,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{system}{Character "76" or "86" defining the isotope plot axis (default
"76").}

\item{Mu1}{Second-stage 238U/204Pb ratio (default 10).}

\item{Kappa}{Second-stage 232Th/238U ratio (default 4).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}
}
\value{
A ggplot2 layer object.
}
\description{
This Geom is used for drawing and labeling isochron, geochron, and kappa
lines used for isotope age model referencing used in lead isotope biplots.
The lines follows the model used by Stacey and Kramers (1975).
}
\details{
The plotting system follows the convention of showing geochron and isochron
lines for the 207Pb/204Pb vs. 206Pb/204Pb plot and the kappa lines for
208Pb/204Pb vs. 206Pb/204Pb plot.
}
\section{Note}{
Currently the plot will scale the xlim and ylim to their maximum bounds. To
prevent this use \code{\link[ggplot2:coord_cartesian]{coord_cartesian(xlim, ylim)}}
to force the axis range to the desiered values.
}

\section{Additional parameters}{

\itemize{
\item \code{Ti}  Initial time of the second stage in years (default 3.7 Ga).
\item \code{interval} Time interval for isochron labels in years (default 200 Ma).
\item \code{show_geochron} Logical; should the Geochron be plotted? (default \code{TRUE}).
\item \code{show_isochrons} Logical; should time isochrons be plotted? (default \code{TRUE})
\item \code{kappa_list} Numeric vector of Kappa values to plot in "86" system.
}
}

\examples{
# example code

library(ggplot2)
set.seed(50)
df <- data.frame(
  pb64 = rnorm(10, 18,0.2),
  pb74 = rnorm(10, 15.7,0.1),
  pb84 = rnorm(10, 37.5,0.1)
)

# Creating the Pb 207/204~206/204 plot
ggplot(df, aes(x = pb64, y = pb74)) +
  geom_point() +
  geom_sk_lines(system = "76") +
  geom_sk_labels(system = "76") +
  coord_cartesian(
    xlim = range(df$pb64) * c(.99, 1.01),
    ylim = range(df$pb74) * c(.99, 1.01)
  ) +
  labs(
    x = expression({}^206*Pb / {}^204*Pb),
    y = expression({}^207*Pb / {}^204*Pb),
  )

# Creating the Pb 208/204~206/204 plot
ggplot(df, aes(x = pb64, y = pb84)) +
  geom_point() +
  geom_sk_lines(system = "86",
                show_isochrons = FALSE, show_geochron = FALSE,
                kappa_list = c(3.2, 3.4, 3.6, 3.8)) +
  geom_sk_labels(system = "86",
                 show_isochrons = FALSE, show_geochron = FALSE,
                 kappa_list = c(3.2, 3.4, 3.6, 3.8)) +
  coord_cartesian(
    xlim = range(df$pb64) * c(.99, 1.01),
    ylim = range(df$pb84) * c(.99, 1.01)
  ) +
  labs(
    x = expression({}^206*Pb / {}^204*Pb),
    y = expression({}^207*Pb / {}^204*Pb),
  )

# Creating the Pb 207/204~206/204 plot with a seperate Geocrone color

ggplot(df, aes(x = pb64, y = pb74)) +
  geom_point() +
  geom_sk_lines(system = "76", show_geochron = FALSE) +
  geom_sk_lines(system = "76", show_isochrons = FALSE,
                color = 'red', linetype = 'dashed') +
  geom_sk_labels(system = "76", show_geochron  = FALSE) +
  geom_sk_labels(system = "76", show_isochrons = FALSE, color = 'red') +
  coord_cartesian(
    xlim = range(df$pb64) * c(.99, 1.01),
    ylim = range(df$pb74) * c(.99, 1.01)
  ) +
  labs(
    x = expression({}^206*Pb / {}^204*Pb),
    y = expression({}^207*Pb / {}^204*Pb),
  )

}
\references{
Stacey, J.S. and Kramers, J.D. (1975) Approximation of
terrestrial lead isotope evolution by a two-stage model. Earth and
Planetary Science Letters 26(2), pp. 207–221.
\url{https://dx.doi.org/10.1016/0012-821X(75)90088-6}.
}
\seealso{
Other Pb isotope functions: 
\code{\link{age_models}}
}
\concept{Pb isotope functions}
\section{Aesthetics}{

\code{geom_path()} understands the following aesthetics. Required aesthetics are displayed in bold and defaults are displayed for optional aesthetics:
\tabular{rll}{
 • \tab \strong{\code{\link[ggplot2:aes_position]{x}}} \tab   \cr
 • \tab \strong{\code{\link[ggplot2:aes_position]{y}}} \tab   \cr
 • \tab \code{\link[ggplot2:aes_colour_fill_alpha]{alpha}} \tab → \code{NA} \cr
 • \tab \code{\link[ggplot2:aes_colour_fill_alpha]{colour}} \tab → via \code{theme()} \cr
 • \tab \code{\link[ggplot2:aes_group_order]{group}} \tab → inferred \cr
 • \tab \code{\link[ggplot2:aes_linetype_size_shape]{linetype}} \tab → via \code{theme()} \cr
 • \tab \code{\link[ggplot2:aes_linetype_size_shape]{linewidth}} \tab → via \code{theme()} \cr
}

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
