% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_geom_spider.R
\name{geom_spider}
\alias{geom_spider}
\title{Spidergram geom for ggplot2}
\usage{
geom_spider(
  mapping = NULL,
  data = NULL,
  inherit.aes = TRUE,
  elements = ifelse(inherits(data, "ASTR"), get_concentration_columns(data), NULL),
  reference = NULL,
  na.rm = FALSE,
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{elements}{Character vector with the list of elements to be plotted. If
data supplied to the geom are in an \link[=ASTR]{ASTR object}, the default value
plots all chemical elements.}

\item{reference}{Character string with what the data should be normalised to;
see \link{normalise_data} fur further details. If \code{NULL}, the default, data will
not be normalised.}

\item{na.rm}{Logical value to indicate whether \code{NA} should be removed from
the data. The default \code{FALSE} will supply all data.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are often
aesthetics used to set a fixed value, such as \code{colour = "red"} or \code{alpha = 0.5}.}
}
\description{
This geom creates spidergrams, line graphs of element concentrations. The
geom supports data normalisation, which is commonly done before plotting.
}
\details{
This geom is special because no x and y coordinates are provided in the
input. Instead, the aesthetic \code{elements} must be provided only in
the\link[ggplot2:aes]{ggplot2} function.

The elements can be supplied either as user-defined character vector or
pre-made sets. See \link{standard_groups} for a list of available sets and the
examples below for how to include them. The examples below also show how to
avoid the default sorting into alphabetical order. The geom will
throw an error if any of the supplied elements is not matched by a column
name in the supplied data.
}
\section{Aesthetics}{
 \code{geom_spidergram()} understands the following aesthetic
values (required aesthetics are in bold): ### update as needed
\itemize{
\item \strong{\code{elements}} (the list of elements to be plotted)
\item \code{alpha}
\item \code{colour} (controls the polygon outline)
\item \code{fill} (controls the polygon fill)
\item \code{linetype}
\item \code{linewidth} (controls the outline thickness)
\item \code{text.colour} (controls colour of the text)
\item \code{size.unit} (controls size aesthetic in "mm", "pt", "cm", "in" or "pc")

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.
}
}

\examples{
# include example with data[[standard_groups$REE]]

library(ggplot2)

test <- data.frame(
  Sample = c("A","B"),
  Yb = c(8,9),
  La = c(10,5),
  Ce = c(20,8)
)

ggplot(test) + geom_spider(mapping = aes(x = standard_groups$REE, color = Sample))

}
