% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_spider.R
\name{geom_spider}
\alias{geom_spider}
\title{Spidergram geom for ggplot2}
\usage{
geom_spider(
  mapping = NULL,
  data = NULL,
  inherit.aes = TRUE,
  reference = NULL,
  na.rm = FALSE,
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{reference}{\code{NULL}, the default, if data should not be normalised or
name of the geochemical reference composition to which data should be
normalised. See \link{references_geochem} for a list of names.}

\item{na.rm}{Logical: remove NA values}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are often
aesthetics used to set a fixed value, such as \code{colour = "red"} or \code{alpha = 0.5}.}
}
\description{
Descriptions goes here ###################
\subsection{This will likely only work with ASTR objects or similarly wide data frame!?}{
}

\subsection{Or is there a way to identify short vs. long form!?}{
}

\subsection{e.g. single value to x and col content = character: long format, x = character vector length > 1 and column content of x = numeric: wide format}{
}

\subsection{y = single value and numeric: long format, concentration, y is not provided: wide format}{
}

\subsection{long format values do not need to be transformed (but normalised!?), wide format must be transformed to long format}{
}
}
\details{
Normalisation is done with \code{\link[=normalise_geochem]{normalise_geochem()}}. If data are geochemically
normalised, only normalised elements are plotted. Otherwise, all elements
are plotted.
}
\examples{
# include example with data[[standard_groups$REE]]

library(ggplot2)

test <- data.frame(
  Sample = c("A","B"),
  Yb = c(8,9),
  La = c(10,5),
  Ce = c(20,8)
)

ggplot(test) + geom_spider(mapping = aes(x = standard_groups$REE, color = Sample))

}
