% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{get_30s_mean}
\alias{get_30s_mean}
\title{Calculate 30 second averages for select variables}
\usage{
get_30s_mean(df)
}
\arguments{
\item{df}{Pass a UPAS v2 or v2+ log data frame from 'read_ast_log' function.}
}
\value{
A modified data frame with 30 second mean data for select variables.
}
\description{
Calculate 30 second averages for select variables
}
\examples{
upasv2x_30s_mean <- df_30s_mean(upasv2x_log)
}
