% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tz_string}
\alias{get_tz_string}
\title{Get the time zone string to be used to format local times in an Access Sensor Technologies (AST) air sampler log file}
\usage{
get_tz_string(UTCOffset, tz = NA)
}
\arguments{
\item{UTCOffset}{A numeric value from the log file header representing the
offset, in hours, between the UTC time and the local time zone associated
with the sample. In UPAS log files, this variable is named "GPSUTCOffset".
In  Home Health Box v2 log files, this variable is named "UTCOffset".}

\item{tz}{Optional: A character string specifying the tz database time zone
that should be used to display local times. Example tz database time zones
include: "America/New_York", "America/Denver", and "America/Los_Angeles".
For additional information, see: \url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}

If the GPSUTCOffset in your log file is a fraction of an hour, it's
best to specify the optional \code{tz} parameter; otherwise, local times in the log file
will be incorrectly labeled as UTC time.}
}
\value{
A string containing the name of a time zone or an \code{NA} value.
}
\description{
\code{get_tz_string()} returns a string representing the time zone with which
local times in a log file should be formatted. If the GPSUTCOffset in your
log file is a whole number of hour, \code{get_tz_string()} function will return a
string. If the GPSUTCOffset in your log file is a fraction of an hour, it's
best to specify the optional \code{tz} parameter to for easier time stamp handling;
otherwise, \code{get_tz_string()}will return \code{NA}.
}
\examples{
# Using a standard UPASv2x log file
upasv2x_filename <- 'PSP00270_LOG_2024-07-10T19_30_20UTC_StartOnNext_____----------.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_header <- read_ast_header(upasv2x_file, update_names=FALSE)
get_tz_string(upasv2x_header$GPSUTCOffset)
}
