% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{gps_map}
\alias{gps_map}
\title{Generate a gps map from a data frame with time-averaged data}
\usage{
gps_map(df)
}
\arguments{
\item{df}{Pass a UPAS v2+ log data frame from 'get_30s_mean' function.}
}
\value{
A leaflet map of 30s averages of the selected variable.
}
\description{
Generate a gps map from a data frame with time-averaged data
}
