% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_map.R
\name{gps_map}
\alias{gps_map}
\title{Generate a gps map from an Access Sensor Technologies (AST) UPAS v2.1 PLUS log data frame}
\usage{
gps_map(df, variable = c("PM2_5MC", "CO2"))
}
\arguments{
\item{df}{A log data frame returned by the \link{read_ast_log} function
with the argument \code{update_names = TRUE}.}

\item{variable}{A column name from the log file. Choose "PM2_5MC" for PM2.5,
and "CO2" for CO2 data.}
}
\value{
A gps map of 30 second averaged data for the selected variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{gps_map} generates a gps map of 30 second averaged PM2.5 or CO2 data from
an Access Sensor Technologies
UPAS v2.1 PLUS log data frame.
If UPASv2 data is provided, a map will not be generated for that device.
This function calculates a 30 second average for any data logged
at a less than 30 second log interval.
The UPAS must have been outside with GPS reception and collecting PM2.5 or CO2 data.
}
\examples{
multiple_upas_logs <- system.file("extdata", package = "astr", mustWork = TRUE) |>
    list.files(pattern="^PS.*.txt$", full.names = TRUE) \%>\%
    lapply(read_ast_log, update_names=TRUE) \%>\%
    dplyr::bind_rows()

upasv2x_pm25_map <- gps_map(multiple_upas_logs, variable="PM2_5MC")
upasv2x_CO2_map <- gps_map(multiple_upas_logs, variable="CO2")
}
