% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_isoprov_dist.R
\name{isoprov_dist}
\alias{isoprov_dist}
\title{Euclidean Distance for Pb Isotope rations to ore Sources}
\usage{
isoprov_dist(
  x,
  ref,
  dist_type = stop("Distance type should be Defined 'ed' or 'mfd'"),
  .n = 1,
  s = 0.001,
  ...
)
}
\arguments{
\item{x}{Matrix, data frame of Pb Isotopes with columns in the order of
206Pb/204Pb, 207Pb/204Pb,208Pb/204Pb or a \code{liaendmember} object.}

\item{ref}{\code{ref.data} object used for Reference analysis.}

\item{dist_type}{Distance type to use, simple euclidean ('ed') or
mass-fractionation corrected ('mfd')}

\item{.n}{Length of result output (Default = 1)}

\item{s}{Mass-fractionation factor (Default = 0.001)}

\item{...}{Additional params}
}
\value{
List of data frame or character vector
}
\description{
Calculate the euclidean distance of each isotope sample to a reference dataset,
and gives the closest regions to the groups.
Mass-fractionation follows the procedure outlined in Albarede et.al (2024)
}
\examples{

# Create data structure

col_names <- c("206Pb/204Pb", "207Pb/204Pb", "208Pb/204Pb")
df <- ASTR::ArgentinaDatabase[
ASTR::ArgentinaDatabase$`Mining area` == "Distrito El Aguilar",
col_names]
ref_data <- as.ref_data(ASTR::ArgentinaDatabase, col_names, "Mining area")

# Create object with class liaendmembers
end_members <- endmembers(
        df,
        col_names,
        tolerance = c(0.01, 0.01),
        clamp = c(Inf, Inf)
)
# Print summary of the liaendmembers object
summary(end_members)
# Get Uclidian distance
isoprov_dist(end_members, ref_data, dist_type = "ed")
# Get Mass Fractionnation Corrected distance
isoprov_dist(end_members, ref_data, dist_type = "mfd")
# Train XGBOOST model on Reference data
\dontrun{ml_model <- train_data(ref_data)}
# Get XGBOOST predicted ml_resutls
isoprov_predict(end_members, ml_model)
}
\references{
Albarede, F., Davis, G., Blichert-Toft, J., Gentelli, L., Gitler, H., Pinto, M., & Telouk, P. (2024).
A new algorithm for using Pb isotopes to determine the provenance of bullion in ancient Greek coinage.
Journal of Archaeological Science, 163, 105919. https://doi.org/10.1016/j.jas.2023.105919
}
