% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isotopes_data}
\alias{isotopes_data}
\title{Isotopes}
\format{
a vector
}
\usage{
isotopes_data
}
\description{
List of naturally occurring isotopes, retrieved from https://www.ciaaw.org/isotopic-abundances.htm,
sorted according to chemical element and isotope number
}
\seealso{
Other chemical_reference_data: 
\code{\link{conversion_oxides}},
\code{\link{elements_data}},
\code{\link{oxides_data}},
\code{\link{special_oxide_states}}
}
\concept{chemical_reference_data}
\keyword{datasets}
