% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_data.R
\docType{data}
\name{ml_model}
\alias{ml_model}
\title{Machine Learning model for lead isotope provenaces}
\format{
A list of xgb models, with length equal to number of groups.
}
\source{
\url{https://globalid.dmt-lb.de/}
}
\usage{
ml_model
}
\description{
Lead isotope data model based on GlobaLID for machine learning provenance.
}
\keyword{datasets}
