% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_normalise_wrapper.R
\name{normalise_data}
\alias{normalise_data}
\title{family
Data normaliation}
\usage{
normalise_data(df, ...)
}
\arguments{
\item{...}{}
}
\description{
family
Data normaliation
}
\seealso{
Other Data normalisation: 
\code{\link{normalise_geochem}()}
}
\concept{Data normalisation}
