% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_normalise_geochem.R
\name{normalise_geochem}
\alias{normalise_geochem}
\title{Geochemical normalisation of data}
\usage{
normalise_geochem(df, reference = names(references_geochem))
}
\arguments{
\item{df}{A data frame in wide format.}

\item{reference}{Character string with the normalisation. See Details for
further information.}
}
\value{
The data frame with the selected elements normalised.
}
\description{
Normalises selected elemental concentrations to a reference composition such
as chondrite or MORB.
}
\details{
See \link{references_geochem} for the supported reference compositions. Additional
reference compositions can be added as named vectors, see examples. They
won't be stored permanently in the package. If you would like to include a
new reference composition, , please reach out to the package maintainers or
create a pull request in the \href{https://github.com/archaeothommy/ASTR}{package's GitHub repo} to add it.

The function converts all elements in the dataframe for which a reference
composition is available. A warning informs if not all elements in the
reference composition are included in the data frame, and an error if none is
included.
}
\examples{
df <- data.frame(
  Sample = c("A","B"),
  La = c(10,5),
  Ce = c(20,8)
)

normalise_geochem(
  df,
  elements = c("La","Ce"),
  reference = "chondrite"
)

}
