% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_geochem.R
\name{normalise_geochem}
\alias{normalise_geochem}
\title{Geochemical normalisation of data}
\usage{
normalise_geochem(df, reference = names(references_geochem))
}
\arguments{
\item{df}{A data frame in wide format.}

\item{reference}{Character string with the normalisation. See Details for
further information.}
}
\value{
The data frame with the selected elements normalised.
}
\description{
Normalises selected elemental concentrations to a reference composition such
as chondrite or MORB.
}
\details{
See \link{references_geochem} for supported geochemical reference compositions and
which elements are covered by which reference composition. The function
converts all elements in the dataframe for which a reference composition is
available. A warning informs if not all elements in the reference composition
are included in the data frame, and an error if none is included.
}
\examples{
df <- data.frame(
  Sample = c("A","B"),
  La = c(10,5),
  Ce = c(20,8)
)

normalise_spider(
  df,
  elements = c("La","Ce"),
  reference = "chondrite"
)

}
\references{
Sun, S.-s. & McDonough, W.F. (1989). Chemical and isotopic
systematics of oceanic basalts. Geological Society, London, Special
Publications 42, 313-345. Table 1, page 318.
}
