% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxides_data}
\alias{oxides_data}
\title{Oxides}
\format{
a vector
}
\usage{
oxides_data
}
\description{
List of oxides, sorted according to alphabet.
}
\seealso{
Other chemical_reference_data: 
\code{\link{conversion_oxides}},
\code{\link{elements_data}},
\code{\link{isotopes_data}},
\code{\link{special_oxide_states}}
}
\concept{chemical_reference_data}
\keyword{datasets}
