% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_PbIso_endmembers.R
\name{pb_iso_endmembers}
\alias{pb_iso_endmembers}
\title{Find Endmembers from a list of LIA points}
\usage{
pb_iso_endmembers(
  x,
  col = NULL,
  tolerance = c(0.01, 0.01),
  clamp = c(Inf, Inf),
  ...
)
}
\arguments{
\item{x}{data.frame or matrix object containing
Pb 206/204, 207/204, 208/204 isotope ratios.}

\item{col}{Isotope column names containing Pb
206/204, 207/204, 208/204 isotope ratios.
Names must contains the significant numbers 6, 7 and 8.}

\item{tolerance}{Vector of length two, with corresponding group 1 and group
2 tolerance value for points considered to be intercepted.
(Default c(0.01, 0.01))}

\item{clamp}{Limit filter for points away from the principle component end
based on Euclidean distance, (Default c(Inf, Inf))}

\item{...}{Additional Parameters}
}
\value{
An object of class 'pbisoendmembers' as a list of 6
\item{data}{Isotope data}
\item{pca_ends}{Endmembers maped using PCA}
\item{group1,group2}{End member groups}
\item{mixing}{Mixing Group}
\item{tolerance}{Totlarance value}
\item{clamp}{Clamping values}
\item{pca}{List of PCA analysis}
}
\description{
Finds endmembers from a set of
Lead Isotope Points using Principle Component
analysis and the Geochron slope according to the two-stage model.
}
\examples{

# Create data structure

col_names <- c("206Pb/204Pb", "207Pb/204Pb", "208Pb/204Pb")
df <- ASTR::ArgentinaDatabase[
ASTR::ArgentinaDatabase$`Mining area` == "Distrito El Aguilar",
col_names]
ref_data <- as.ref_data(ASTR::ArgentinaDatabase, col_names, "Mining area")

# Create object with class liaendmembers
end_members <- pb_iso_endmembers(
        df,
        col_names,
        tolerance = c(0.01, 0.01),
        clamp = c(Inf, Inf)
)
# Print summary of the liaendmembers object
summary(end_members)
# Get Uclidian distance
pb_iso_prov_dist(end_members, ref_data, dist_type = "ed")
# Get Mass Fractionnation Corrected distance
pb_iso_prov_dist(end_members, ref_data, dist_type = "mfd")
# Train XGBOOST model on Reference data
\dontrun{ml_model <- pb_iso_train_data(ref_data)}
# Get XGBOOST predicted ml_resutls
pb_iso_prov_predict(end_members, ml_model)
}
