% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_PbIso_Prov.R
\name{pb_iso_prov_dist}
\alias{pb_iso_prov_dist}
\alias{pb_iso_prov_dist.ASTR}
\alias{euc_dist}
\alias{euc_dist.ASTR}
\alias{mf_dist}
\alias{mf_dist.ASTR}
\title{Euclidean distance for Pb isotope ratios to ore sources}
\usage{
pb_iso_prov_dist(x, ...)

\method{pb_iso_prov_dist}{ASTR}(
  x,
  ref,
  ref_group,
  dist_type = c("ed", "mf", "all"),
  .n = 1,
  s = 0.001,
  ...
)

euc_dist(x, ...)

\method{euc_dist}{ASTR}(x, ref, ref_group, .n = 1, ...)

mf_dist(x, ...)

\method{mf_dist}{ASTR}(x, ref, ref_group, .n = 1, s = 0.001, ...)
}
\arguments{
\item{x}{ASTR object containing 206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb isotope
ratios for analysis.}

\item{...}{Additional parameters.}

\item{ref}{ASTR object containing 206Pb/204Pb, 207Pb/204Pb, 208Pb/204Pb
isotope ratios as a provenance reference.}

\item{ref_group}{Name of the column containing isotope groups as a character
string.}

\item{dist_type}{Distance type to use, simple Euclidean (\code{"ed"}) or
mass-fractionation corrected (\code{"mfd"}).}

\item{.n}{Length of result output, default value is \code{1}.}

\item{s}{Mass fractionation factor, default value is \code{0.001}.}
}
\value{
ASTR object of matched x, reference, and corresponding distance
values.
}
\description{
Calculates the Euclidean distance of each isotope sample to a reference data
set and gives the closest regions to the groups. Mass fractionation follows
the procedure outlined in Albarede et al. (2024).
}
\examples{
GlobaLID_ASTR_ref <- as_pbiso_ref_data(GlobaLID_ASTR,
                                   names(GlobaLID_ASTR)[[2]],
                                   min_groupsize = 5)

# Euclidean distance
euc_dist(tel_dor, GlobaLID_ASTR_ref, .n = 1)

# Mass fractionation correction
mf_dist(tel_dor, GlobaLID_ASTR_ref, .n = 1, s = 0.001)

# Wrapper function where the reference data is of class 'ASTR_Pb_iso_ref_data'
pb_iso_prov_dist(tel_dor,
                 GlobaLID_ASTR,
                 ref_group = names(GlobaLID_ASTR)[[2]],
                 dist_type = "all")

# Wrapper function where reference data is not of class 'ASTR_Pb_iso_ref_data'
pb_iso_prov_dist(tel_dor,
                 GlobaLID_ASTR_ref,
                 dist_type = "all")

}
\references{
Albarede, F., Davis, G., Blichert-Toft, J., Gentelli, L., Gitler,
H., Pinto, M., and Telouk, P. (2024) A new algorithm for using Pb isotopes
to determine the provenance of bullion in ancient Greek coinage. Journal of
Archaeological Science 163, 105919.
\url{https://doi.org/10.1016/j.jas.2023.105919}
}
\seealso{
Other Pb isotope functions:
\code{\link{age_models}},
\code{\link[=as_pbiso_ref_data]{as_pbiso_ref_data()}},
\code{\link[=geom_sk_lines]{geom_sk_lines()}},
\code{\link[=pb_iso_endmembers]{pb_iso_endmembers()}},
\code{\link[=pb_iso_prov_predict]{pb_iso_prov_predict()}},
\code{\link[=pb_iso_train_data]{pb_iso_train_data()}},
\code{\link{tel_dor}}
}
\concept{Pb isotope functions}
