% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_PbIso_Prov.R
\name{pb_iso_prov_predict}
\alias{pb_iso_prov_predict}
\alias{pb_iso_prov_predict.ASTR}
\title{Predict isotope provenance}
\usage{
pb_iso_prov_predict(x, ...)

\method{pb_iso_prov_predict}{ASTR}(x, model_list = NULL, .top = 1, ...)
}
\arguments{
\item{x}{Matrix of \code{pb_iso_endmembers} object of pb isotope samples.}

\item{...}{Additional parameters.}

\item{model_list}{Model list generated by \code{train_data()}.}

\item{.top}{Number of highest probability values considered for final output,
default value is \code{1}.}
}
\value{
Data frame object or list of data frames.
}
\description{
Predicts Pb isotope provenance of a sample matrix using a XGBoost trained
list
}
\seealso{
pb_iso_train_data

Other Pb isotope functions:
\code{\link{age_models}},
\code{\link[=as_pbiso_ref_data]{as_pbiso_ref_data()}},
\code{\link[=geom_sk_lines]{geom_sk_lines()}},
\code{\link[=pb_iso_endmembers]{pb_iso_endmembers()}},
\code{\link[=pb_iso_prov_dist]{pb_iso_prov_dist()}},
\code{\link[=pb_iso_train_data]{pb_iso_train_data()}},
\code{\link{tel_dor}}
}
\concept{Pb isotope functions}
