% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_PbIso_Prov.R
\name{pb_iso_prov_predict}
\alias{pb_iso_prov_predict}
\alias{pb_iso_prov_predict.ASTR}
\title{Predict Isotope Provenance}
\usage{
pb_iso_prov_predict(x, ...)

\method{pb_iso_prov_predict}{ASTR}(x, model_list = NULL, .top = 1, ...)
}
\arguments{
\item{x}{Matrix of \code{pbisoendmembers} object of pbisotope samples}

\item{...}{Additional params}

\item{model_list}{Model list generated by \code{train_data()}}

\item{.top}{Number of highest probality values considred for final output (Default = 1)}
}
\value{
data.frame object or list of data.frames
}
\description{
Predicts Pb Isotope provenance of a sample matrix in reference to a xgboost trained list
}
\seealso{
train_data

Other Pb isotope functions:
\code{\link{age_models}},
\code{\link[=as_pbiso_ref_data]{as_pbiso_ref_data()}},
\code{\link{data}},
\code{\link[=geom_sk_lines]{geom_sk_lines()}},
\code{\link[=pb_iso_endmembers]{pb_iso_endmembers()}},
\code{\link[=pb_iso_prov_dist]{pb_iso_prov_dist()}},
\code{\link[=pb_iso_train_data]{pb_iso_train_data()}}
}
\concept{Pb isotope functions}
