% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{read_ast_header}
\alias{read_ast_header}
\title{Read the header data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
read_ast_header(file, update_names = FALSE)
}
\arguments{
\item{file}{Any AST air sampler  log file name.}

\item{update_names}{Option to update old sampler names to latest version.}
}
\value{
A data frame with header data in wide format.
}
\description{
Read the header data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
upasv2_filename <- 'PS0166_LOG_2021-09-29T17_37_09UTC_test_______________---.txt'
upasv2_file <- system.file("extdata", upasv2_filename, package = "astr", mustWork = TRUE)
upasv2_header <- read_ast_header(upasv2_file, update_names=FALSE)
upasv2x_filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_header <- read_ast_header(upasv2x_file, update_names=FALSE)
}
