% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{read_ast_header}
\alias{read_ast_header}
\title{Read the header data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
read_ast_header(file, update_names = FALSE, shiny = FALSE)
}
\arguments{
\item{file}{Any AST air sampler  log file name.}

\item{update_names}{Option to update old sampler names to latest version.}

\item{shiny}{Option to make TRUE if using function with AST shiny app.}
}
\value{
A data frame with header data in wide format.
}
\description{
Read the header data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
upasv2_filename <- 'PS0166_LOG_2021-09-29T17_37_09UTC_test_______________---.txt'
upasv2_file <- system.file("extdata", upasv2_filename, package = "astr", mustWork = TRUE)
upasv2_header <- read_ast_header(upasv2_file, update_names=FALSE)
upasv2x_filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
upasv2x_filename <- 'PSP00030_LOG_2022-05-11T23_24_01UTC_---------------_----------.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_header <- read_ast_header(upasv2x_file, update_names=FALSE)
upasv2x_diag_filename <- 'PSP00055_LOG_2022-03-24T18_05_32UTC_DIAGNOSTIC________________.txt'
upasv2x_diag_filename <- 'PSP00030_LOG_2022-09-29T17_04_19UTC_DIAGNOSTIC________________.txt'
upasv2x_diag_file <- system.file("extdata", upasv2x_diag_filename, package = "astr", mustWork = TRUE)
upasv2x_diag_header <- read_ast_header(upasv2x_diag_file, update_names=FALSE)
upasv2x_rev136_filename <- 'PSP00110_LOG_2022-09-27T23_51_30UTC_BatteryTest_____Test______.txt'
upasv2x_rev136_filename <- 'PSP00110_LOG_2022-09-24T23_08_33UTC_FirmwareTest____8P3_______.txt'
upasv2x_rev136_file <- system.file("extdata", upasv2x_rev136_filename, package = "astr", mustWork = TRUE)
upasv2x_rev136_header <- read_ast_header(upasv2x_rev136_file, update_names=FALSE)
}
