% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{read_ast_log}
\alias{read_ast_log}
\title{Read the full log data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
read_ast_log(
  file,
  tz_offset = NA,
  update_names = FALSE,
  cols_keep = c(),
  cols_drop = c(),
  shiny = FALSE
)
}
\arguments{
\item{file}{Any AST air sampler log file name.}

\item{tz_offset}{Pass an option timezone offset.}

\item{update_names}{Option to update old sampler names to latest version.}

\item{cols_keep}{Specify log file columns to keep.}

\item{cols_drop}{Specify log file columns to remove.}

\item{shiny}{Option to make TRUE if using function with AST shiny app.}
}
\value{
A data frame with all log data plus some header data appended.
}
\description{
Read the full log data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
upasv2_filename <- 'PS0166_LOG_2021-09-29T17_37_09UTC_test_______________---.txt'
upasv2_file <- system.file("extdata", upasv2_filename, package = "astr", mustWork = TRUE)
upasv2_log <- read_ast_log(upasv2_file)
upasv2x_filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_log <- read_ast_log(upasv2x_file)
filename2 <- 'SH00007_LOG_2021-12-13T13_28_41UTC_---------------_-----.txt'
file2 <- system.file("extdata", filename2, package = "astr", mustWork = TRUE)
data_ast_log <- read_ast_log(file2)
data_ast_log <- read_ast_log(file2, cols_keep = c("SampleTime","UnixTime","DateTimeUTC","DateTimeLocal","PM2_5MC"))
upasv2x_rev136_filename <- 'PSP00110_LOG_2022-09-27T23_51_30UTC_BatteryTest_____Test______.txt'
upasv2x_rev136_file <- system.file("extdata", upasv2x_rev136_filename, package = "astr", mustWork = TRUE)
upasv2x_rev136_log <- read_ast_log(upasv2x_rev136_file, update_names=FALSE)
}
