% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{read_ast_log}
\alias{read_ast_log}
\title{Read the full log data from an Access Sensor Technologies (AST) air sampler
log file}
\usage{
read_ast_log(file, tz_offset = NA)
}
\arguments{
\item{file}{Any AST air sampler  log file name.}

\item{tz_offset}{Pass an option timezone offset}
}
\value{
A data frame with all log data plus some header data appended.
}
\description{
Read the full log data from an Access Sensor Technologies (AST) air sampler
log file
}
\examples{
filename <- 'PSP00024_LOG_2021-08-11T18_18_03UTC_test____________test______.txt'
filename2 <- 'SH00007_LOG_2021-12-13T13_28_41UTC_---------------_-----.txt'
file <- system.file("extdata", filename, package = "astr", mustWork = TRUE)
data_ast_log <- read_ast_log(file)
file2 <- system.file("extdata", filename2, package = "astr", mustWork = TRUE)
data_ast_log <- read_ast_log(file2)
}
