% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{references_geochem}
\alias{references_geochem}
\title{Geochemical reference compositions}
\format{
An object of class \code{list} of length 2.
}
\usage{
references_geochem
}
\description{
List with reference values for common geochemical normalisation such as chondrite or MORB.
}
\references{
Sun, S.-s. & McDonough, W.F. (1989). Chemical and isotopic systematics
}
\keyword{datasets}
