% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_axis}
\alias{shiny_axis}
\title{Rename UPAS log file single column name
to be more user friendly for the Shiny app plot}
\usage{
shiny_axis(clm_name)
}
\arguments{
\item{df}{Pass a UPAS v2 or v2+ log column name from a formatted data frame.}

\item{df_h}{Pass a UPAS v2 or v2+ header data frame from 'read_ast_log' function.}
}
\value{
A modified column name for plot axis label in shinyAST app.
}
\description{
Rename UPAS log file single column name
to be more user friendly for the Shiny app plot
}
\examples{
plot_label <- shiny_axis(column_name)
}
