% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_header}
\alias{shiny_header}
\title{Rename and format Access Sensor Technologies (AST) UPAS v2 or UPAS v2.1 PLUS
header data frame columns and add units
for the online shinyAST app}
\usage{
shiny_header(df_h, fract_units = FALSE)
}
\arguments{
\item{df_h}{A header data frame returned by the \link{read_ast_header} function
with the argument \code{update_names = TRUE}.}

\item{fract_units}{Boolean to specify if units should be fractional (L min^-1 vs L/min).}
}
\value{
A modified single row header data frame with column names and units
for the online shinyAST app.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{shiny_header} modifies an Access Sensor Technologies UPAS v2 or UPAS v2.1 PLUS
header data frame
for use with the online shinyAST data analysis app.
This function adds units to the applicable variable names.
}
\examples{
multiple_upas_headers <- list.files(path = "inst/extdata", pattern="^PS.*.txt$",
                                  full.names = TRUE) \%>\%
        lapply(read_ast_header, update_names = TRUE) \%>\%
        dplyr::bind_rows()

upas_shiny_header <- shiny_header(multiple_upas_headers)
}
