% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_log}
\alias{shiny_log}
\title{Rename UPAS log file data frame columns
to be more user friendly for the Shiny app}
\usage{
shiny_log(df)
}
\arguments{
\item{df}{Pass a upasv2x dataframe from read_ast_header function.}
}
\value{
A modified data frame with units added to column names of log data.
}
\description{
Rename UPAS log file data frame columns
to be more user friendly for the Shiny app
}
\examples{
upasv2x_log_units <- units_upasv2x_log(upasv2x_log)
}
