% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_log}
\alias{shiny_log}
\title{Format Access Sensor Technologies (AST) UPAS v2 or UPAS v2.1 PLUS
log data frame columns
for the online shinyAST app}
\usage{
shiny_log(df)
}
\arguments{
\item{df}{A log data frame returned by the \link{read_ast_log} function
with the argument \code{update_names = TRUE}.}
}
\value{
A modified log data frame with extraneous variables removed,
more used variables brought to the front of the data frame,
and SampleTime formatted to hours for the online shinyAST app plots.
}
\description{
\code{shiny_log} modifies an Access Sensor Technologies UPAS v2 or UPAS v2.1 PLUS
log data frame
for use with the online shinyAST data analysis app.
This function removes and reorders some columns to create a better user interface.
}
\examples{
multiple_upas_logs <- list.files(path = "inst/extdata", pattern="^PS.*.txt$",
    full.names = TRUE) \%>\%
    lapply(read_ast_log, update_names=TRUE) \%>\%
    dplyr::bind_rows()

upas_shiny_log <- shiny_log(multiple_upas_logs)
}
