% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ast.R
\name{shiny_units}
\alias{shiny_units}
\title{Reformat units for shinyAST app}
\usage{
shiny_units(vect)
}
\arguments{
\item{vect}{Pass a vector.}
}
\value{
A vector with fractional units instead of the
standard UPASv2 and UPASv2+ log file unit format
}
\description{
Reformat units for shinyAST app
}
\examples{
colnames(upasv2x_header_shiny) <- shiny_units(colnames(upasv2x_header_shiny)
clm_name <- shiny_units(clm_name)
}
