% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bracketing.R
\name{standard_sample_bracketing}
\alias{standard_sample_bracketing}
\title{Standard Sample Bracketing}
\usage{
standard_sample_bracketing(data, ID_std = "", header = "", pos = 0, mp = 1000)
}
\arguments{
\item{data}{< Dataframe with analysis results from the machine>}

\item{ID_std}{string, \if{html}{\out{<ID of the standard for bracketing>}}. The default value is "".}

\item{pos}{integer \if{html}{\out{<position of the first line for bracketing>}}.}

\item{mp}{integer \if{html}{\out{<multiplier>}}.}
}
\value{
A data frame
}
\description{
Performs Standard Sample Bracketing (SSB) correction for isotope ratio data.
}
\details{
This function corrects for instrumental drift or mass bias during isotopic analyses.
The measured isotope value of each sample is linearly interpolated between the two standard measurements taken immediately before and after the sample measurement.
This function is a useful tool for automating a common practice in handling isotope data. It reduces the likelihood of human error and saves time.
}
\examples{
<write here any example code. These are small simple examples on how to use
the function or to highlight specific features>
}
\references{
Mason, Thomas F.D., Dominik J. Weiss, Matthew Horstwood, Randall R. Parrish, Sara S. Russell, Eta Mullaneb and Barry J. Colesa. 2004. High-precision Cu and Zn isotope analysis by plasma source mass spectrometry. Journal of Analytical Atomic Spectrometry, 19, 209-217. DOI	https://doi.org/10.1039/B306958C
}
