% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oxide_conversion.R
\name{sum_duplicates}
\alias{sum_duplicates}
\title{Sum duplicate columns (e.g., Fe from FeO and Fe2O3)}
\usage{
sum_duplicates(mat)
}
\description{
Sum duplicate columns (e.g., Fe from FeO and Fe2O3)
}
\keyword{internal}
