% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASTR_utils.R
\name{transform_notation}
\alias{transform_notation}
\title{Transform unit in ratio notation into notation with negative exponents}
\usage{
transform_notation(unit)
}
\arguments{
\item{unit}{Character string}
}
\description{
Transform unit in ratio notation into notation with negative exponents
}
\keyword{internal}
