% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_header.R
\name{transpose_ast_header}
\alias{transpose_ast_header}
\title{Transpose the header data from an Access Sensor Technologies air sampler log
file and format the transposed header as a data frame.}
\usage{
transpose_ast_header(header, diag = NULL)
}
\arguments{
\item{header}{A data table containing header data read using the
\link{fread_ast_header} function.}

\item{diag}{An optional data table containing diagnostic test data read using
the \link{fread_ast_header} function. If the diag argument is specified, key
metrics from the diagnostic test will be added to the header data frame.}
}
\value{
A data frame of header data.
}
\description{
\code{transpose_ast_header} Takes the data read using the \link{fread_ast_header}
function and transposes those data so each variable in the header is a unique
column in a data frame.
}
\examples{
# UPASv2 EXAMPLES
upasv2_filename <- 'PS1771_LOG_2024-06-13T21_20_17UTC_GPSoutside_________Eng.txt'
upasv2_file <- system.file("extdata", upasv2_filename, package = "astr", mustWork = TRUE)
upasv2_header_list <- fread_ast_header(upasv2_file)
upasv2_header_wide <- transpose_ast_header(upasv2_header_list$header)

# UPASv2x EXAMPLES
upasv2x_filename <- 'PSP00270_LOG_2024-06-25T21_37_48UTC_GPS-in-out______----------.txt'
upasv2x_file <- system.file("extdata", upasv2x_filename, package = "astr", mustWork = TRUE)
upasv2x_header_list <- fread_ast_header(upasv2x_file)
upasv2x_header_wide <- transpose_ast_header(upasv2x_header_list$header)

# Diagnostic file
upasv2x_diag_filename <- 'PSP00270_LOG_2024-06-13T16_24_47UTC_DIAGNOSTIC________________.txt'
upasv2x_diag_file <- system.file("extdata", upasv2x_diag_filename, package = "astr", mustWork = TRUE)
upasv2x_diag_header_list <- fread_ast_header(upasv2x_diag_file)
upasv2x_diag_header_wide <- transpose_ast_header(upasv2x_diag_header_list$header, upasv2x_diag_header_list$diag)

# HHB EXAMPLES
hhb_filename <- 'HHB00032_LOG_2024-07-01T18_20UTC.csv'
hhb_file <- system.file("extdata", hhb_filename, package = "astr", mustWork = TRUE)
hhb_header_raw <- fread_ast_header(hhb_file)$header
hhb_header_wide <- transpose_ast_header(hhb_header_raw)
}
