% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2x.R
\name{upasv2x_sample_settings}
\alias{upasv2x_sample_settings}
\title{Create sample settings dataframe from an Access Sensor Technologies (AST)
air sampler header dataframe}
\usage{
upasv2x_sample_settings(df_h, shiny = FALSE, fract_units = FALSE)
}
\arguments{
\item{df_h}{Pass a upasv2x header dataframe from read_ast_header function.}
}
\value{
A modified data frame with only the sample settings data.
}
\description{
Create sample settings dataframe from an Access Sensor Technologies (AST)
air sampler header dataframe
}
\examples{
upasv2x_sample_settings <- upasv2x_sample_settings(upasv2x_header)
}
