% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upasv2x.R
\name{upasv2x_sample_summary}
\alias{upasv2x_sample_summary}
\title{Create sample summary dataframe from an Access Sensor Technologies (AST)
air sampler header dataframe}
\usage{
upasv2x_sample_summary(df_h, df = NULL)
}
\arguments{
\item{df_h}{Pass a upasv2x header dataframe from read_ast_header function.}
}
\value{
A modified data frame with only the sample summary data.
}
\description{
Create sample summary dataframe from an Access Sensor Technologies (AST)
air sampler header dataframe
}
\examples{
upasv2x_sample_summary <- upas2x_sample_summary(upasv2x_header, upasv2x_log)
upasv2x_sample_summary <- upas2x_sample_summary(upasv2x_header)
}
